% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-dates.R
\name{as_period_date}
\alias{as_period_date}
\title{Convert dates series into periodic dates series}
\usage{
as_period_date(
  dates_series,
  period = c("day", "month", "quarter", "year"),
  period_date = c("start", "end")
)
}
\arguments{
\item{dates_series}{A vector of date/timestamps.}

\item{period}{A character of period, e.g. "day", "month",
"quarter", "year". Default "day".}

\item{period_date}{A character of period_date format, e.g. "start",
"end", "start" format date as start of the period, "end" format date as end
of period. Default "start".}
}
\value{
A date_series with period format.
}
\description{
Convert dates series into periodic dates series
}
\seealso{
Other utils_dates: 
\code{\link{guess_dates_period}()},
\code{\link{is_periodic_dates}()}
}
\concept{utils_dates}
