% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicator-generate.R
\name{backup_indicators}
\alias{backup_indicators}
\title{Backup indicators in batch mode}
\usage{
backup_indicators(stock_db, ds_indicator_defs, backup_dir = "backup")
}
\arguments{
\item{stock_db}{A stock database object to operate.}

\item{ds_indicator_defs}{A dataframe of indicator definition to delete.}

\item{backup_dir}{A path of dir to save backup indicators. Default
"backup" is the subdirectory of indicator dir which is specified by
\code{dir_path_db(stock_db, "DIR_DB_DATA_INDICATOR")}}
}
\value{
Path of backup dir. Raise error if anything goes wrong.
}
\description{
Backup customized indicators of stock_db in batch into backup dir.
}
\seealso{
Other indicator generate functions: 
\code{\link{delete_indicators}()},
\code{\link{generate_indicators}()}
}
\concept{indicator generate functions}
