% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor-test.R
\name{build_sort_portfolios}
\alias{build_sort_portfolios}
\title{Build sort portfoilos basing on group list of stocks}
\usage{
build_sort_portfolios(stocks_list, factor_value_list, ngroup = 3,
  first_group_index = 1, factor_group_order = c("desc", "asc"))
}
\arguments{
\item{stocks_list}{a list of stkcds of stocks.}

\item{factor_value_list}{a list of factor value of corresponding stocks.}

\item{ngroup}{numbers of groups to cut stocks into, default 3.}

\item{first_group_index}{index number of group name, default 1 for group_1.}

\item{factor_group_order}{order of factor group_order:
\itemize{
\item asc:  group_1(smallest factor) ... group_N(largest factor)
\item desc: group_1(largest factor)  ... group_N(smallest factor)
}}
}
\value{
a tibble dataset of portfolio_group, stkcd, weight,
                        factor_value.
}
\description{
use stocks list and factor value list to build sort portfoilos by sorting
factor value and cutting in n groups.
}
