% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-fields.R
\name{check_fields}
\alias{check_fields}
\title{Check whether fields exsit in data or not.}
\usage{
check_fields(data, fields)
}
\arguments{
\item{data}{a dataframe or matrix to check.}

\item{fields}{a character vector of field names to check.}
}
\value{
return invisble NULL if succeed, otherwise raise error.
}
\description{
Check whether fields exsit in data or not.
}
\seealso{
Other utils_fields: \code{\link{expect_type_fields}},
  \code{\link{is_type_field}}
}
\concept{utils_fields}
