% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor-normalize.R
\name{clean_extremes_sigma}
\alias{clean_extremes_sigma}
\title{Clean extremes by normal sigma method}
\usage{
clean_extremes_sigma(x, n_sigma = 3, extreme_value = c("limit", "NA"))
}
\arguments{
\item{x}{a vector of data.}

\item{n_sigma}{numbers of sigma to identify extemes, by default 3.}

\item{extreme_value}{value to replace extremes, i.e. "limit","NA",
by default "limit".}
}
\value{
a vector of data without extremes
}
\description{
Identify extremes in data vector by normal sigma method, and repalce extremes
by NA or max/min limitation
}
