% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-mgt.R, R/data-mgt-gta.R
\docType{methods}
\name{clear_tables}
\alias{clear_tables}
\alias{clear_tables,gta_db-method}
\title{Clear data in all tables in stock database}
\usage{
clear_tables(stock_db, ...)

\S4method{clear_tables}{gta_db}(stock_db, ...)
}
\arguments{
\item{stock_db}{A stock database object to operate.}
}
\value{

}
\description{
Generic function to Clear existed data in all tables in stock_db.
}
\section{Methods (by class)}{
\itemize{
\item \code{gta_db}: clear data in all tables in a database of gta_db class
}}

\seealso{
Other data managment functions: \code{\link{get_datasource}},
  \code{\link{import_table}}
}
\concept{data managment functions}
