% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock-db.R, R/stock-db-gta.R
\docType{methods}
\name{close_stock_db}
\alias{close_stock_db}
\alias{close_stock_db,gta_db-method}
\title{Close the stock database}
\usage{
close_stock_db(stock_db, ...)

\S4method{close_stock_db}{gta_db}(stock_db, ...)
}
\arguments{
\item{stock_db}{a stock database object to operate}
}
\value{
TRUE if success, else FALSE
}
\description{
Generic function to close a stock database
}
\section{Methods (by class)}{
\itemize{
\item \code{gta_db}: Close a database of gta_db class
}}

\examples{
stock_db <- stock_db(gta_db, "GTA_SQLData")
open_stock_db(stock_db)
close_stock_db(stock_db)
}
