% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock-db.R, R/stock-db-gta.R
\docType{methods}
\name{code2name}
\alias{code2name}
\alias{code2name,gta_db-method}
\title{Translate code into name}
\usage{
code2name(x, code, ...)

\S4method{code2name}{gta_db}(x, code, ...)
}
\arguments{
\item{x}{a object containg code/name infomation}

\item{code}{a code or a vector of codes to be translated}

\item{...}{other arguments to be provided to underlyling functions}
}
\value{
a name or a vector of names
}
\description{
Generic function to translate code into name
}
\section{Methods (by class)}{
\itemize{
\item \code{gta_db}: Translate code into name in a database of gta_db class
}}

