% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock-db.R
\name{code_name_list}
\alias{code_name_list}
\title{Class creator of code_name_list class}
\usage{
code_name_list(codes, names)
}
\arguments{
\item{codes}{a vector of code}

\item{names}{a vector of name}
}
\value{
a object of code_name_list
}
\description{
code_name_list class creator
}
\examples{

}
