% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor-composite.R
\name{composite_factor}
\alias{composite_factor}
\title{Build composite factor by aggregating related factors}
\usage{
composite_factor(ds_factors, aggregate_formula, drop = TRUE)
}
\arguments{
\item{ds_factors}{a factors dataset.}

\item{aggregate_formula}{a formula to aggregate factors into new factor,
e.g. \code{composite_factor ~ 0.33*PE + 0.33*PB + 0.33*PS}.}

\item{drop}{a logic to determine whether to drop original factors in the
result datasets, by default TRUE.}
}
\value{
\preformatted{       a dataset of new factor.
}
}
\description{
Composite factor is built by aggregating formula coming from weight of
similar or correlated factors, which is determined by various methods, such as ....
}
\details{
How to build composition
}
