% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicator-build.R
\name{compute_indicator}
\alias{compute_indicator}
\title{Compute Indicator by Using Definition and Variables}
\usage{
compute_indicator(ts_vars, ind_def_fun, ...,
  date_index_field = c("date"), key_fields = NULL, parallel = TRUE)
}
\arguments{
\item{ts_vars}{A dataframe of variable timeseries to compute indicator.}

\item{ind_def_fun}{A function of defining indicator.}

\item{...}{Params to ind_def_fun.}

\item{date_index_field}{Name of date index field of ts_vars, default 'date'.}

\item{key_fields}{A character vector of key fields, which identify unique
observation in each date.}

\item{parallel}{A logic to deterimine whether to use parallel processing.
default TRUE means to use parallel processing.}
}
\value{
A dataframe of indicator timeseries if succeed, otherwise NULL.
}
\description{
Use definition function and variables timeseries to compute indicator.
}
\seealso{
Other indicator build functions: \code{\link{create_indicator_def}},
  \code{\link{get_indicator_defs}},
  \code{\link{get_indicator_vars}},
  \code{\link{parse_indicator_vars}}
}
\concept{indicator build functions}
