% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicator-define.R
\name{create_attribute_def_fun}
\alias{create_attribute_def_fun}
\title{Create definition function of new attribute for modifying indicator}
\usage{
create_attribute_def_fun(attr_name, attr_fun, ...)
}
\arguments{
\item{attr_name}{A character for attribute name.}

\item{attr_fun}{A function to generate attribute.}

\item{...}{Other arguments to attr_fun.}
}
\value{
A function of attribute definition to modify indicator if succeed,
otherwise NULL.
}
\description{
Create a definition function of new attribute for modifying indicator.
The definition function is used by \code{\link{modify_indicator}} to compute
attribute in modifying existed indicator timeseries of stocks.
}
\seealso{
Other indicator define functions: 
\code{\link{create_indicator_def_fun}()},
\code{\link{get_indicator_defs}()},
\code{\link{get_indicator_vars}()},
\code{\link{ind_attr_def_indcd}()},
\code{\link{ind_attr_def_trdstat}()},
\code{\link{parse_indicator_vars}()},
\code{\link{prioritize_indicator_defs}()}
}
\concept{indicator define functions}
