% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicator-define.R
\name{create_indicator_def}
\alias{create_indicator_def}
\title{Translate indicator params into definiton function}
\usage{
create_indicator_def(indicator_expr, indicator_name, rolly_window = 0,
  period = c("day", "month", "quarter", "year"))
}
\arguments{
\item{indicator_expr}{A expr as a formula to compute indicator.}

\item{indicator_name}{A character for indicator name.}

\item{rolly_window}{A integer as rolly computing window.}

\item{period}{A periodicity of indicator, i.e. "day", "month",
"quarter", "yearly".}
}
\value{
A function of indicator definition to compute indicator if succeed,
otherwise NULL.
}
\description{
Combine indicator params into a definition function for computing indicators.
}
\seealso{
Other indicator build functions: \code{\link{compute_indicator}},
  \code{\link{get_indicator_defs}},
  \code{\link{get_indicator_vars}},
  \code{\link{parse_indicator_vars}}
}
\concept{indicator build functions}
