% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicator-generate.R
\name{delete_indicators}
\alias{delete_indicators}
\title{Delete indicators in batch mode}
\usage{
delete_indicators(stock_db, ds_indicator_defs)
}
\arguments{
\item{stock_db}{A stock database object to operate.}

\item{ds_indicator_defs}{A dataframe of indicator definition to delete.}
}
\value{
NULL invisibly. Raise error if anything goes wrong.
}
\description{
Delete customized indicators of stock_db in batch, which clear all customized
built indicator files.
}
\seealso{
Other indicator generate functions: 
\code{\link{backup_indicators}()},
\code{\link{generate_indicators}()}
}
\concept{indicator generate functions}
