% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock-db-gta.R
\name{dir_path_db.gta_db}
\alias{dir_path_db.gta_db}
\title{gta_db class}
\usage{
dir_path_db.gta_db(stock_db, dir_id = c("DIR_DB_DATA",
  "DIR_DB_DATA_SOURCE", "DIR_DB_DATA_ORIGIN", "DIR_DB_DATA_LOG",
  "DIR_DB_DATA_INDICATOR"), force = TRUE)
}
\description{
gta_db class
}
