% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock-db.R, R/stock-db-gta.R
\docType{methods}
\name{dir_path_db}
\alias{dir_path_db}
\alias{dir_path_db,gta_db-method}
\title{Get Path of Data Directory from stock_db}
\usage{
dir_path_db(stock_db, dir_id = c("DIR_DB_DATA", "DIR_DB_DATA_SOURCE",
  "DIR_DB_DATA_ORIGIN", "DIR_DB_DATA_LOG", "DIR_DB_DATA_INDICATOR"),
  force = TRUE, ...)

\S4method{dir_path_db}{gta_db}(stock_db, dir_id = c("DIR_DB_DATA",
  "DIR_DB_DATA_SOURCE", "DIR_DB_DATA_ORIGIN", "DIR_DB_DATA_LOG",
  "DIR_DB_DATA_INDICATOR"), force = TRUE, ...)
}
\arguments{
\item{stock_db}{A stock database object to operate.}

\item{dir_id}{A character id of directory.
\itemize{
      \item DIR_DB_DATA: dir of database of stock_db;
      \item DIR_DB_DATA_SOURCE : dir of source data to process into database;
      \item DIR_DB_DATA_ORIGIN:  dir of origin data to import into database;
      \item DIR_DB_DATA_LOG: dir of log of database operation;
      \item DIR_DB_DATA_INDICATOR: dir of customized indicators.
      }}

\item{force}{Whether return result if dir dosen't existed.
Default TRUE, return result if dir doesn't exist.}
}
\value{
a full path of dir if succeed, otherwise NULL.
 If the path of dir dosen't exist and force = FALSE, it will raise a error.
}
\description{
Generic function to get path of data directory from stock_db.
}
\section{Methods (by class)}{
\itemize{
\item \code{gta_db}: get path of data directory from a database of
gta_db class
}}

\seealso{
Other stock_db generics: \code{\link{close_stock_db}},
  \code{\link{code2name}},
  \code{\link{fetch_table_dataset}},
  \code{\link{get_factors_info}},
  \code{\link{get_factors}},
  \code{\link{get_finacial_report}},
  \code{\link{get_indicators_from_source}},
  \code{\link{get_indicators}},
  \code{\link{get_market_return}},
  \code{\link{get_profile}},
  \code{\link{get_riskfree_rate}},
  \code{\link{get_spt_stocks}},
  \code{\link{get_stock_dataset}},
  \code{\link{get_stock_industry}},
  \code{\link{get_stock_return}},
  \code{\link{get_table_dataset}},
  \code{\link{init_stock_db}},
  \code{\link{list_stock_tables}}, \code{\link{name2code}},
  \code{\link{open_stock_db}},
  \code{\link{save_indicators_to_source}}
}
\concept{stock_db generics}
