% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-testthat-helpers.R
\name{expect_not_null}
\alias{expect_not_null}
\title{Extend testthat expectation for not null}
\usage{
expect_not_null(object, info = NULL, label = NULL)
}
\arguments{
\item{object}{Object to test.}

\item{info}{Extra information to be included in the message.
This argument is soft-deprecated and should not be used in new code.
Instead see alternatives in \code{\link[testthat:quasi_label]{testthat::quasi_label()}}.}

\item{label}{Used to customize failure messages. For expert use only.}
}
\description{
This is an expectation as a counterpart of \code{\link[testthat:expect_null]{testthat::expect_null()}}.
}
\examples{
\dontrun{

x <- NULL
y <- 10

expect_not_null(y)

testthat::show_failure(expect_not_null(x))
}
}
