% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-fields.R
\name{expect_type_fields}
\alias{expect_type_fields}
\title{Identify fields with specified type}
\usage{
expect_type_fields(
  data,
  expect_type = c("numeric", "integer", "double", "character", "date", "factor",
    "list", "NA"),
  negate = FALSE,
  predicate_fun = NULL,
  ...
)
}
\arguments{
\item{data}{A vector of date/timestamps.}

\item{expect_type}{A character of type to test, e.g. "numeric", "integer",
"double", "character", "date", "factor", "list", "NA"}

\item{negate}{A logic flag of whether to negate expect_type of not,
Default FALSE means
not to negate the expect_type.}

\item{predicate_fun}{A function used as testing, return TRUE if matched.
Default NULL means not to use predicate_fun for testing.}

\item{...}{Params to predicate_fun.}
}
\value{
A vectors of field names with specified type. If no field is
expect type, return a character(0).
}
\description{
Get the field names of dataframe with specified type.
}
\seealso{
Other utils_fields: 
\code{\link{is_type_field}()},
\code{\link{verify_fields}()}
}
\concept{utils_fields}
