% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-factor-explore.R
\name{exploreFactorsCorrelation}
\alias{exploreFactorsCorrelation}
\title{Explore factors correlation Server Function}
\usage{
exploreFactorsCorrelation(input, output, session, ds_factors)
}
\arguments{
\item{input}{input from caller.}

\item{output}{output to caller.}

\item{session}{Session from caller.}

\item{ds_factors}{A data.frame of factors.}
}
\description{
Explore factors correlation Server Function
}
