% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-factor-explore.R
\name{exploreFactorsCorrelationUI}
\alias{exploreFactorsCorrelationUI}
\title{Explore factors correlation UI}
\usage{
exploreFactorsCorrelationUI(id, ds_factors)
}
\arguments{
\item{id}{An ID string of the UI.}

\item{ds_factors}{A data.frame of factors.}
}
\description{
Explore factors correlation UI
}
