% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-factor-explore.R
\name{exploreFactorsDistribution}
\alias{exploreFactorsDistribution}
\title{Explore factors distribution Server Function}
\usage{
exploreFactorsDistribution(input, output, session, ds_factors)
}
\arguments{
\item{input}{input from caller.}

\item{output}{output to caller.}

\item{session}{Session from caller.}

\item{ds_factors}{A data.frame of factors.}
}
\description{
Explore factors distribution Server Function
}
