% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-factor-explore.R
\name{exploreFactorsDistributionUI}
\alias{exploreFactorsDistributionUI}
\title{Explore factors distribution UI}
\usage{
exploreFactorsDistributionUI(id, ds_factors)
}
\arguments{
\item{id}{An ID string of the UI.}

\item{ds_factors}{A data.frame of factors.}
}
\description{
Explore factors distribution UI
}
