% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor-test.R
\name{factor_test_IC}
\alias{factor_test_IC}
\title{Creator of factor_test_IC class}
\usage{
factor_test_IC(
  ds_test,
  IC_fun,
  ...,
  output_type = c("summary", "raw"),
  factor_field = "factor_name",
  date_field = "date"
)
}
\arguments{
\item{ds_test}{A timeseries dataset with descriptors of factors for test.}

\item{IC_fun}{A function to compute information coefficients.}

\item{...}{Arguments passed to IC_fun.}

\item{output_type}{Type of output data, i.e."summary", "raw", if "raw",
raw data will be append to output object for diagnosis.}

\item{factor_field}{Name of factor field of ds_test, by default "factor_name".}

\item{date_field}{Name of date field of ds_test, by default "date",
Column must be date-like.}
}
\value{
\preformatted{            A object of factor_test_IC class.
}
}
\description{
Conduct information coefficients test for descriptors of factors and build object
of factor_test_IC class as output.
}
