% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor-composite.R
\name{factors_weight_equal}
\alias{factors_weight_equal}
\title{Compute factors weight for compositing factor by equal-weight method}
\usage{
factors_weight_equal(factors_list)
}
\arguments{
\item{factors_list}{a list of factors.}
}
\value{
\preformatted{ a list of factorS weight with columns of factor_name, factor_weight.
}
}
\description{
Compute factors weight by giving each factor equal weight of
\eqn{1/number of factors}. The result of factors weight could be used for
\code{\link{weight_formula}} to build a formula of composite factor
}
