% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock-db.R, R/stock-db-gta.R
\name{fetch_table_dataset}
\alias{fetch_table_dataset}
\alias{fetch_table_dataset,gta_db-method}
\title{Fetch many datasets from stock_db}
\usage{
fetch_table_dataset(stock_db, table_list, ...)

\S4method{fetch_table_dataset}{gta_db}(stock_db, table_list, ...)
}
\arguments{
\item{stock_db}{A stock database object to operate.}

\item{table_list}{A character vector of table names.}

\item{...}{Extra arguments to be passed to methods.}
}
\value{
A list of names of table fetched if succeed.
}
\description{
Generic function to fetch many datasets from stock_db.
}
\section{Functions}{
\itemize{
\item \code{fetch_table_dataset,gta_db-method}: get many datasets from a database of gta_db class
}}

\seealso{
Other stock_db generics: 
\code{\link{close_stock_db}()},
\code{\link{code2name}()},
\code{\link{dir_path_db}()},
\code{\link{get_factors_info}()},
\code{\link{get_factors}()},
\code{\link{get_financial_report}()},
\code{\link{get_indicators_from_source}()},
\code{\link{get_indicators}()},
\code{\link{get_market_return}()},
\code{\link{get_profile}()},
\code{\link{get_riskfree_rate}()},
\code{\link{get_spt_stocks}()},
\code{\link{get_stock_dataset}()},
\code{\link{get_stock_industry}()},
\code{\link{get_stock_info}()},
\code{\link{get_stock_return}()},
\code{\link{get_table_dataset}()},
\code{\link{init_stock_db}()},
\code{\link{list_stock_tables}()},
\code{\link{name2code}()},
\code{\link{open_stock_db}()},
\code{\link{save_indicators_to_source}()}
}
\concept{stock_db generics}
