% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicator-generate.R
\name{generate_indicators}
\alias{generate_indicators}
\title{Generate indicators in batch mode}
\usage{
generate_indicators(
  stock_db,
  ds_indicator_defs,
  validate_def = FALSE,
  validate_stkcds = c("600031", "000157", "600066", "000550"),
  parallel = TRUE,
  log_file_prefix = "generate_indicator_log",
  log_dir = "./log"
)
}
\arguments{
\item{stock_db}{A stock database object to operate.}

\item{ds_indicator_defs}{A dataframe of indicator definition to generate.}

\item{validate_def}{A logic flag to determine whether to validate indicator
definition or not.Default FALSE, means to produce indicators on full
dataset, TRUE means to validate definition on small datasets.}

\item{validate_stkcds}{A character vector of stock codes used for validating
indicator definition.}

\item{parallel}{A logic to determine whether to use parallel processing.
Default TRUE means to use parallel processing.}

\item{log_file_prefix}{A character of log file prefix to name log file. Log
file is named as format of "log_file_prefix(current).csv" Default is
"generate_indicator_log".}

\item{log_dir}{Path to save log file. If NULL, don't save log file by default
"./log".}
}
\value{
NULL invisibly. Raise error if anything goes wrong.
}
\description{
Generate customized indicators of stock_db in batch, and save into files.
}
\details{
There are two methods of generating indicators:
\itemize{
\item \strong{formal production}: produce final result of indicators on full dataset
from stock_db, which will take longer time ;
\item \strong{validating definition}: produce trial results of indicators on small
dataset from stock_db, which will take shorter time.
}
}
\seealso{
Other indicator generate functions: 
\code{\link{backup_indicators}()},
\code{\link{delete_indicators}()}
}
\concept{indicator generate functions}
