% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock-db.R
\name{get_assets_return}
\alias{get_assets_return}
\title{Get assets return from market and stocks return}
\usage{
get_assets_return(benchmark_return, stocks_return)
}
\arguments{
\item{benchmark_return}{A timeseries of market index}

\item{stocks_return}{A timeseries of a group stock stocks}
}
\value{
A timeseries of assets return
}
\description{
Get assets return timeseries by combining marekt return anad stocks return
}
