% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock-db.R
\name{get_factor_indicator}
\alias{get_factor_indicator}
\title{Get factor indicator timeseries from stock_db}
\usage{
get_factor_indicator(stock_db, factor_list, ...)
}
\arguments{
\item{stock_db}{A stock database object to operate.}

\item{factor_list}{factor name list.}
}
\value{
A timeseries of factor indicator
}
\description{
Generic function to get  factor indicator from stock_db.
}
\seealso{
Other stock_db generics: \code{\link{close_stock_db}},
  \code{\link{code2name}},
  \code{\link{fetch_table_dataset}},
  \code{\link{get_factors_info}},
  \code{\link{get_market_return}},
  \code{\link{get_stock_dataset}},
  \code{\link{get_stock_return}},
  \code{\link{get_table_dataset}},
  \code{\link{init_stock_db}},
  \code{\link{list_stock_tables}}, \code{\link{name2code}},
  \code{\link{open_stock_db}}
}
