% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock-db.R, R/stock-db-gta.R
\docType{methods}
\name{get_factors}
\alias{get_factors}
\alias{get_factors,gta_db-method}
\title{Get factors from stock_db}
\usage{
get_factors(stock_db, factor_codes, ...)

\S4method{get_factors}{gta_db}(stock_db, factor_codes, ...)
}
\arguments{
\item{stock_db}{A stock database object to operate.}

\item{factor_codes}{A vector of factor code.}
}
\value{
A dataframe of factor timeseries if succeed, otherwise NULL.
  In order to store multi-factors with different periods(day,
  month, quarter, year), returned data frame adapts longer format,
  which means name of factor is stored in "factor_name", value of factor
  is stored in "factor_value".
}
\description{
Generic function to get factor timeseries from stock_db.
}
\section{Methods (by class)}{
\itemize{
\item \code{gta_db}: get factor timeseries from
a database of gta_db class
}}

\seealso{
Other stock_db generics: \code{\link{close_stock_db}},
  \code{\link{code2name}}, \code{\link{dir_path_db}},
  \code{\link{fetch_table_dataset}},
  \code{\link{get_factors_info}},
  \code{\link{get_finacial_report}},
  \code{\link{get_indicators_from_source}},
  \code{\link{get_indicators}},
  \code{\link{get_market_return}},
  \code{\link{get_profile}},
  \code{\link{get_riskfree_rate}},
  \code{\link{get_spt_stocks}},
  \code{\link{get_stock_dataset}},
  \code{\link{get_stock_industry}},
  \code{\link{get_stock_return}},
  \code{\link{get_table_dataset}},
  \code{\link{init_stock_db}},
  \code{\link{list_stock_tables}}, \code{\link{name2code}},
  \code{\link{open_stock_db}},
  \code{\link{save_indicators_to_source}}
}
\concept{stock_db generics}
