% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock-db.R, R/stock-db-gta.R
\docType{methods}
\name{get_finacial_report}
\alias{get_finacial_report}
\alias{get_finacial_report,gta_db-method}
\title{Get financial report timesereis from stock_db}
\usage{
get_finacial_report(stock_db, stock_cd_list = NULL,
  statement = c("blance_sheet", "income", "cashflow_direct",
  "cashflow_indrect", "income_ttm", "cashflow_direct_ttm",
  "cashflow_indrect_ttm"), consolidated = TRUE,
  period_type = c("quarter", "year"), period_date = c("end", "start"),
  ...)

\S4method{get_finacial_report}{gta_db}(stock_db, stock_cd_list = NULL,
  statement = c("blance_sheet", "income", "cashflow_direct",
  "cashflow_indrect", "income_ttm", "cashflow_direct_ttm",
  "cashflow_indrect_ttm"), consolidated = TRUE,
  period_type = c("quarter", "year"), period_date = c("end", "start"),
  ...)
}
\arguments{
\item{stock_db}{A stock database object to operate.}

\item{stock_cd_list}{A character vector of stock cd, default value of NULL means.
all stock data will be returned.}

\item{statement}{A string of statement type, i.e. "blance_sheet",
"income", "cashflow_direct","cashflow_indrect",
"income_ttm", "cashflow_direct_ttm", "cashflow_indrect_ttm".

 @param consolidated A logic inidcate report is consolidated or not.
  Default TRUE means consolidated report, FALSE means parenet company
  report.}

\item{period_type}{Date period for time series, e.g. "quarter",
"year", by default value is  "quarter".}

\item{period_date}{Choose start/end date of period as the date for
observation}
}
\value{
A data frame of report timeseries if succeed, otherwise NULL.
}
\description{
Generic function to get financial report timesereis from stock_db.
}
\section{Methods (by class)}{
\itemize{
\item \code{gta_db}: get financial report timeseries from a database of gta_db class
}}

\seealso{
Other stock_db generics: \code{\link{close_stock_db}},
  \code{\link{code2name}}, \code{\link{dir_path_db}},
  \code{\link{fetch_table_dataset}},
  \code{\link{get_factors_info}},
  \code{\link{get_factors}},
  \code{\link{get_indicators_from_source}},
  \code{\link{get_indicators}},
  \code{\link{get_market_return}},
  \code{\link{get_profile}},
  \code{\link{get_riskfree_rate}},
  \code{\link{get_spt_stocks}},
  \code{\link{get_stock_dataset}},
  \code{\link{get_stock_industry}},
  \code{\link{get_stock_return}},
  \code{\link{get_table_dataset}},
  \code{\link{init_stock_db}},
  \code{\link{list_stock_tables}}, \code{\link{name2code}},
  \code{\link{open_stock_db}},
  \code{\link{save_indicators_to_source}}
}
\concept{stock_db generics}
