% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicator-define.R, R/indicator-define-gta.R
\docType{methods}
\name{get_indicator_defs}
\alias{get_indicator_defs}
\alias{get_indicator_defs,gta_db-method}
\title{Get definitions of customzied indicators from stock_db}
\usage{
get_indicator_defs(stock_db, ...)

\S4method{get_indicator_defs}{gta_db}(stock_db, ...)
}
\arguments{
\item{stock_db}{A stock database object to operate.}
}
\value{
A dataframe of definitions of customized indicators if succeed,
otherwise NULL.
}
\description{
Generic function to get definition of customized indicators from stock_db.
}
\section{Methods (by class)}{
\itemize{
\item \code{gta_db}: get indicator defs from database
of gta_db class.
}}

\seealso{
Other indicator build functions: \code{\link{compute_indicator}},
  \code{\link{create_indicator_def}},
  \code{\link{get_indicator_vars}},
  \code{\link{parse_indicator_vars}}
}
\concept{indicator build functions}
