% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicator-define-gta.R
\name{get_indicator_vars.gta_db}
\alias{get_indicator_vars.gta_db}
\title{indicators from database of gta_db class.}
\usage{
get_indicator_vars.gta_db(stock_db, indicator_defs)
}
\description{
indicators from database of gta_db class.
}
