% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicator-define.R, R/indicator-define-gta.R
\docType{methods}
\name{get_indicator_vars}
\alias{get_indicator_vars}
\alias{get_indicator_vars,gta_db-method}
\title{Get input vars for computing customized indicators from stock_db}
\usage{
get_indicator_vars(stock_db, indicator_defs, ...)

\S4method{get_indicator_vars}{gta_db}(stock_db, indicator_defs, ...)
}
\arguments{
\item{stock_db}{A stock database object to operate.}

\item{indicator_defs}{A dataframe of indicator definitions, which require
vars to compute indicators.}
}
\value{
A dataframe of input vars for computing customized indicators
if succeed, otherwise NULL.
}
\description{
Generic function to get vars for computing customized indicators
from stock_db.
}
\section{Methods (by class)}{
\itemize{
\item \code{gta_db}: get input vars for computing customzied
indicators from database of gta_db class.
}}

\seealso{
Other indicator define functions: \code{\link{create_attribute_def_fun}},
  \code{\link{create_indicator_def_fun}},
  \code{\link{get_indicator_defs}},
  \code{\link{ind_attr_def_indcd}},
  \code{\link{ind_attr_def_trdstat}},
  \code{\link{parse_indicator_vars}},
  \code{\link{prioritize_indicator_defs}}
}
\concept{indicator define functions}
