% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock-db.R, R/stock-db-gta.R
\name{get_indicators}
\alias{get_indicators}
\alias{get_indicators,gta_db-method}
\title{Get indicators from stock_db}
\usage{
get_indicators(stock_db, indicator_codes, ...)

\S4method{get_indicators}{gta_db}(stock_db, indicator_codes, ...)
}
\arguments{
\item{stock_db}{A stock database object to operate.}

\item{indicator_codes}{A vector of indicator code.}

\item{...}{Extra arguments to be passed to methods.}
}
\value{
A dataframe of indicator timeseries if succeed, otherwise NULL.
In order to store multi-indicators with different periods(day,
month, quarter, year), returned data frame adapts longer format,
which means code of indicator is stored in "ind_code", value of factor
is stored in "ind_value".
}
\description{
Generic function to get indicator timeseries from stock_db.
}
\section{Functions}{
\itemize{
\item \code{get_indicators,gta_db-method}: get indicator timeseries from
a database of gta_db class
}}

\seealso{
Other stock_db generics: 
\code{\link{close_stock_db}()},
\code{\link{code2name}()},
\code{\link{dir_path_db}()},
\code{\link{fetch_table_dataset}()},
\code{\link{get_factors_info}()},
\code{\link{get_factors}()},
\code{\link{get_financial_report}()},
\code{\link{get_indicators_from_source}()},
\code{\link{get_market_return}()},
\code{\link{get_profile}()},
\code{\link{get_riskfree_rate}()},
\code{\link{get_spt_stocks}()},
\code{\link{get_stock_dataset}()},
\code{\link{get_stock_industry}()},
\code{\link{get_stock_info}()},
\code{\link{get_stock_return}()},
\code{\link{get_table_dataset}()},
\code{\link{init_stock_db}()},
\code{\link{list_stock_tables}()},
\code{\link{name2code}()},
\code{\link{open_stock_db}()},
\code{\link{save_indicators_to_source}()}
}
\concept{stock_db generics}
