% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock-db-gta.R
\name{get_indicators_from_source.gta_db}
\alias{get_indicators_from_source.gta_db}
\title{@param data_filters   A list of filter to apply on result data. The list
 consists of items which has field as its name and filter expr string as
 its content, defualt is list( typrep = "typrep == 'A'",
  markettype = "markettype == 21"). Only valid for gta_db.}
\usage{
get_indicators_from_source.gta_db(stock_db, indicator_source,
  indicator_codes = NULL, ouput_format = c("long", "wide"),
  data_filters = list(typrep = "typrep == 'A'", markettype =
  "markettype == 21"), date_fields = c("date", "accper", "trddt",
  "trdmnt", "trdynt", "clsdt", "shrchdgt"), retain_fields = c("stkcd",
  "indcd"))
}
\arguments{
\item{date_fields}{A character vector of possible date field in original
data, which will be renamed as "date". default is c("date", "accper",
"trddt", "trdmnt", "trdynt","clsdt", "shrchdgt"). Only valid for gta_db.}

\item{retain_fields}{A character vector of possible fields to retain in
final dataset, default is c("stkcd", "indcd"). Only valid for gta_db.}
}
\description{
@param data_filters   A list of filter to apply on result data. The list
 consists of items which has field as its name and filter expr string as
 its content, defualt is list( typrep = "typrep == 'A'",
  markettype = "markettype == 21"). Only valid for gta_db.
}
