% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock-db.R, R/stock-db-gta.R
\docType{methods}
\name{get_indicators_from_source}
\alias{get_indicators_from_source}
\alias{get_indicators_from_source,gta_db-method}
\title{Get indicators from specified source in stock_db}
\usage{
get_indicators_from_source(stock_db, indicator_source,
  indicator_codes = NULL, ouput_format = c("long", "wide"), ...)

\S4method{get_indicators_from_source}{gta_db}(stock_db, indicator_source,
  indicator_codes = NULL, ouput_format = c("long", "wide"), ...)
}
\arguments{
\item{stock_db}{A stock database object to operate.}

\item{indicator_source}{A name of table or file(with extension)
in which indicators are stored. For file, formats of rds,csv are supported.}

\item{indicator_codes}{A vector of inicator code. Default NULL return
all indicators in the source.}

\item{ouput_format}{Output format for data frame, i.e "long" and "wide":
 \itemize{
     \item in long format: name of indicator is stored in "ind_name", value of indicator
is stored in "ind_value";
     \item in wide format: value of indicator is stored in
field named with the name of indicator.
 }}
}
\value{
A data frame of indicator timeseries if succeed, otherwise NULL.
}
\description{
Generic function to get indicator timeseries from specified source in stock_db.
}
\section{Methods (by class)}{
\itemize{
\item \code{gta_db}: get indicator timeseries from
specified source in a database of gta_db class
}}

\seealso{
Other stock_db generics: \code{\link{close_stock_db}},
  \code{\link{code2name}}, \code{\link{dir_path_db}},
  \code{\link{fetch_table_dataset}},
  \code{\link{get_factors_info}},
  \code{\link{get_factors}}, \code{\link{get_indicators}},
  \code{\link{get_market_return}},
  \code{\link{get_profile}},
  \code{\link{get_stock_dataset}},
  \code{\link{get_stock_return}},
  \code{\link{get_table_dataset}},
  \code{\link{init_stock_db}},
  \code{\link{list_stock_tables}}, \code{\link{name2code}},
  \code{\link{open_stock_db}},
  \code{\link{save_indicators_to_source}}
}
\concept{stock_db generics}
