% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock-db.R, R/stock-db-gta.R
\docType{methods}
\name{get_profile}
\alias{get_profile}
\alias{get_profile,gta_db-method}
\title{Get profile of stock_db}
\usage{
get_profile(stock_db, profile_name, ...)

\S4method{get_profile}{gta_db}(stock_db, profile_name, ...)
}
\arguments{
\item{stock_db}{A stock database object to operate.}
}
\value{
full path of profile of stock db if succeed,
otherwise raise a error
}
\description{
Generic function to get path of profile of stock_db.
}
\section{Methods (by class)}{
\itemize{
\item \code{gta_db}: get profile of a database of gta_db class
}}

\seealso{
Other stock_db generics: \code{\link{close_stock_db}},
  \code{\link{code2name}}, \code{\link{dir_path_db}},
  \code{\link{fetch_table_dataset}},
  \code{\link{get_factors_info}},
  \code{\link{get_factors}},
  \code{\link{get_finacial_report}},
  \code{\link{get_indicators_from_source}},
  \code{\link{get_indicators}},
  \code{\link{get_market_return}},
  \code{\link{get_riskfree_rate}},
  \code{\link{get_spt_stocks}},
  \code{\link{get_stock_dataset}},
  \code{\link{get_stock_industry}},
  \code{\link{get_stock_return}},
  \code{\link{get_table_dataset}},
  \code{\link{init_stock_db}},
  \code{\link{list_stock_tables}}, \code{\link{name2code}},
  \code{\link{open_stock_db}},
  \code{\link{save_indicators_to_source}}
}
\concept{stock_db generics}
