% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock-db-gta.R
\name{get_riskfree_rate.gta_db}
\alias{get_riskfree_rate.gta_db}
\title{gta_db class}
\usage{
get_riskfree_rate.gta_db(stock_db, period = c("day", "month", "quarter",
  "year"))
}
\description{
gta_db class
}
