% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock-db.R, R/stock-db-gta.R
\name{get_riskfree_rate}
\alias{get_riskfree_rate}
\alias{get_riskfree_rate,gta_db-method}
\title{Get riskfree rate from stock_db}
\usage{
get_riskfree_rate(stock_db, period = c("day", "month", "quarter", "year"), ...)

\S4method{get_riskfree_rate}{gta_db}(stock_db, period = c("day", "month", "quarter", "year"), ...)
}
\arguments{
\item{stock_db}{A stock database object to operate.}

\item{period}{Date period of time series, e.g. "day", "month",
"quarter", "year", default is "day".}

\item{...}{Extra arguments to be passed to methods.}
}
\value{
A dataframe of riskfree rate timeseries if succeed, otherwise NULL.
}
\description{
Generic function to get riskfree rate timeseries from stock_db.
}
\section{Functions}{
\itemize{
\item \code{get_riskfree_rate,gta_db-method}: get riskfree rate timeseries from a database of
gta_db class
}}

\seealso{
Other stock_db generics: 
\code{\link{close_stock_db}()},
\code{\link{code2name}()},
\code{\link{dir_path_db}()},
\code{\link{fetch_table_dataset}()},
\code{\link{get_factors_info}()},
\code{\link{get_factors}()},
\code{\link{get_financial_report}()},
\code{\link{get_indicators_from_source}()},
\code{\link{get_indicators}()},
\code{\link{get_market_return}()},
\code{\link{get_profile}()},
\code{\link{get_spt_stocks}()},
\code{\link{get_stock_dataset}()},
\code{\link{get_stock_industry}()},
\code{\link{get_stock_return}()},
\code{\link{get_table_dataset}()},
\code{\link{init_stock_db}()},
\code{\link{list_stock_tables}()},
\code{\link{name2code}()},
\code{\link{open_stock_db}()},
\code{\link{save_indicators_to_source}()}
}
\concept{stock_db generics}
