% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock-db.R, R/stock-db-gta.R
\name{get_spt_stocks}
\alias{get_spt_stocks}
\alias{get_spt_stocks,gta_db-method}
\title{Get stock info of special treatment from stock_db}
\usage{
get_spt_stocks(stock_db, ...)

\S4method{get_spt_stocks}{gta_db}(stock_db, ...)
}
\arguments{
\item{stock_db}{A stock database object to operate.}

\item{...}{Extra arguments to be passed to methods.}
}
\value{
A dataframe stock info of special treatment if succeed, otherwise NULL.
}
\description{
Generic function to stock info of special treatment from stock_db.
}
\section{Functions}{
\itemize{
\item \code{get_spt_stocks,gta_db-method}: get stock info of special treatment from
a database of gta_db class
}}

\seealso{
Other stock_db generics: 
\code{\link{close_stock_db}()},
\code{\link{code2name}()},
\code{\link{dir_path_db}()},
\code{\link{fetch_table_dataset}()},
\code{\link{get_factors_info}()},
\code{\link{get_factors}()},
\code{\link{get_financial_report}()},
\code{\link{get_indicators_from_source}()},
\code{\link{get_indicators}()},
\code{\link{get_market_return}()},
\code{\link{get_profile}()},
\code{\link{get_riskfree_rate}()},
\code{\link{get_stock_dataset}()},
\code{\link{get_stock_industry}()},
\code{\link{get_stock_return}()},
\code{\link{get_table_dataset}()},
\code{\link{init_stock_db}()},
\code{\link{list_stock_tables}()},
\code{\link{name2code}()},
\code{\link{open_stock_db}()},
\code{\link{save_indicators_to_source}()}
}
\concept{stock_db generics}
