% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock-db-gta.R
\name{get_stock_dataset.gta_db}
\alias{get_stock_dataset.gta_db}
\title{in a database of gta_db class}
\usage{
get_stock_dataset.gta_db(stock_db, table_name, stock_cd_list = NULL)
}
\description{
in a database of gta_db class
}
