% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock-db-gta.R, R/stock-db.R
\name{get_stock_return.gta_db}
\alias{get_stock_return.gta_db}
\alias{get_stock_return}
\title{Get stock return timeseries from stock_db}
\usage{
\method{get_stock_return}{gta_db}(stock_db, stock_cd_list = NULL,
  period_type = c("daily", "weekly", "monthly", "annual"))

get_stock_return(stock_db, stock_cd_list = NULL, period_type = c("daily",
  "weekly", "monthly", "annual"))
}
\arguments{
\item{stock_db}{a stock database object to operate}

\item{stock_cd_list}{a list of stock cd, default value of NULL means
all stock data will be returned}

\item{period_type}{date peroid for time series, e.g. "daily", "weekly",
"monthly", "annual", default value is "daily"}
}
\value{
A timeseries of stock return
}
\description{
Generic function to get stock return timeseries from stock_db
}
\section{Methods (by class)}{
\itemize{
\item \code{gta_db}: get stock return timeseries from a database of gta_db class
}}

\examples{

}
