% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock-db.R, R/stock-db-gta.R
\docType{methods}
\name{get_table_dataset}
\alias{get_table_dataset}
\alias{get_table_dataset,gta_db-method}
\title{Get adataset from a table in stock_db}
\usage{
get_table_dataset(stock_db, table_name, ...)

\S4method{get_table_dataset}{gta_db}(stock_db, table_name, ...)
}
\arguments{
\item{stock_db}{A stock database object to operate.}

\item{table_name}{Name of target table.}
}
\value{
A data frame on success, or NULL.
}
\description{
Generic function to get one dataset from stock_db.
}
\section{Methods (by class)}{
\itemize{
\item \code{gta_db}: get a table dataset from a database of gta_db class
}}

\examples{
ds_trd_mnth.df <- get_table_dataset(stock_db, table_name = "TRD_Mnth_月个股回报率")
}
\seealso{
Other stock_db generics: \code{\link{close_stock_db}},
  \code{\link{code2name}},
  \code{\link{fetch_table_dataset}},
  \code{\link{get_factor_indicator}},
  \code{\link{get_factors_info}},
  \code{\link{get_market_return}},
  \code{\link{get_stock_dataset}},
  \code{\link{get_stock_return}},
  \code{\link{init_stock_db}},
  \code{\link{list_stock_tables}}, \code{\link{name2code}},
  \code{\link{open_stock_db}}
}
