% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-cliui.R
\name{help_fun}
\alias{help_fun}
\title{Display usage of function}
\usage{
help_fun(fun = NULL, argument_desc = "", examples = "")
}
\arguments{
\item{fun}{a function to be display usage.}

\item{argument_desc}{a character arguments description.}

\item{examples}{a character examples.}
}
\value{
standard output
}
\description{
Display help for a function by arguments usage and examples.
}
\seealso{
Other utils_cliui: 
\code{\link{interactive_call}()},
\code{\link{interactive_fun}()}
}
\concept{utils_cliui}
