% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-mgt.R, R/data-mgt-gta.R
\docType{methods}
\name{import_table}
\alias{import_table}
\alias{import_table,gta_db-method}
\title{Import a raw data file into table in stock database}
\usage{
import_table(stock_db, input_file, input_type = c("csv", "txt"),
  input_dir = NULL, start_index = 2L, target_table = NULL,
  ignore_problems = TRUE, log_dir = "./log", ...)

\S4method{import_table}{gta_db}(stock_db, input_file,
  input_type = c("csv", "txt"), input_dir = NULL, start_index = 2L,
  target_table = NULL, ignore_problems = TRUE, log_dir = "./log",
  ...)
}
\arguments{
\item{stock_db}{A stock database object to operate.}

\item{input_file}{A Name or a path of input data file.}

\item{input_type}{Format of input file, e.g. "txt", "csv".}

\item{input_dir}{Working dir of input file, if NULL, use dir of input file as
working dir, by default NULL,}

\item{start_index}{Start index of first line of actual records in data file,
by default 2L, which means first line is header and actual data starts from
second lines.}

\item{target_table}{Name of target table in stock db, if NULL, use basename
input_file as target table ,  by default NULL,}

\item{ignore_problems}{Whether to ignore problems when covert data,
if TRUE, continue to import data but log problems into log file,
otherwise abort importing process.  By default TRUE.}

\item{log_dir}{Path of log dir for saving problem log file,
by default"./log", if the log path doesn't exsited, it  will be created.}
}
\value{
TRUE if success, else FALSE.
}
\description{
Generic function to import a raw data into target table in stock database.
}
\details{
Import porcess include two phrases:
\enumerate{
   \item Convert raw data into R data type.
   \item Transfer R data into table in database.
}

Any problem in converting raw data will be logged in log file
 in log dir, e.g. "input_file(YYYY-MM-DD).csv"
}
\section{Methods (by class)}{
\itemize{
\item \code{gta_db}: import raw data into table in a database of gta_db class
}}

\seealso{
Other data managment functions: \code{\link{clear_tables}},
  \code{\link{get_datasource}}
}
\concept{data managment functions}
