% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicator-define-gta.R
\name{ind_attr_def_trdstat.gta_db}
\alias{ind_attr_def_trdstat.gta_db}
\title{trading status from database of gta_db class.}
\usage{
ind_attr_def_trdstat.gta_db(stock_db)
}
\description{
trading status from database of gta_db class.
}
