% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicator-define-funs.R
\name{indicator_expr_funs}
\alias{indicator_expr_funs}
\alias{Lag}
\alias{GrowthRate}
\alias{Ratio}
\alias{Demean}
\alias{Quarter_TTM}
\alias{Beta}
\alias{RiskFreeRate}
\title{Utils functions to facilitate defining indicator expr}
\usage{
Lag(x, k = 1)

GrowthRate(x)

Ratio(x_numerator, y_denominator)

Demean(x)

Quarter_TTM(date, x)

Beta(y, x)

RiskFreeRate(stock_db, indicator_code, period = c("day", "month",
  "quarter", "year"))
}
\description{
Functions are only used in defining indicator expr.
}
\details{
In order to facilitate defining indicator expr, there are two kinds of
 expr functions:
 \itemize{
    \item \strong{expr_funs to define customized indicator}:
      calling function to generate a result series, like
       \code{Lag}, \code{GrowthRate}, \code{Ratio},
       \code{Demean}, \code{Quarter_TTM}, \code{Beta}, etc.
    \item \strong{expr_funs to define dynamic indicator}:
      provide dynamic time series from database.

 }
}
\section{Functions}{
\itemize{
\item \code{Lag}: lag a sereis at k period.
(k>0 lag behind, k<0 lag ahead)

\item \code{GrowthRate}: compute growth rate of a sereis.

\item \code{Ratio}: compute ratio of two sereis.

\item \code{Demean}: substract mean from a series.

\item \code{Quarter_TTM}: make a quarterly TTM(Trial Twelve Month)
series.

\item \code{Beta}: compute beta between two varable series

\item \code{RiskFreeRate}: provide timeseries of dynamic indicator of
risfree rate.
}}

\note{
In order to be different form normal R function, we use CamelCase
 stlye for functions name.
}
