% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock-db-gta.R
\name{industry_name_list.gta_db}
\alias{industry_name_list.gta_db}
\title{Create a industry_name_list for a database of gta_db class}
\usage{
industry_name_list.gta_db(stock_db)
}
\description{
Create a industry_name_list for a database of gta_db class
}
