% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock-db-gta.R, R/stock-db.R
\name{init_stock_db.gta_db}
\alias{init_stock_db.gta_db}
\alias{init_stock_db}
\title{Init param of stock db}
\usage{
\method{init_stock_db}{gta_db}(stock_db)

init_stock_db(stock_db)
}
\arguments{
\item{stock_db}{a stock database object to operate}
}
\value{
TRUE if success, else FALSE
}
\description{
Generic function to initiate params of a stock database
}
\section{Methods (by class)}{
\itemize{
\item \code{gta_db}: Init param of database of gta_db class
}}

\examples{
stock_db <- stock_db(gta_db, "GTA_SQLData")
open_stock_db(stock_db)
init_stock_db(stock_db)
}
