% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-cliui.R
\name{interactive_fun}
\alias{interactive_fun}
\title{Turn a function into interactive function}
\usage{
interactive_fun(fun, quiet = FALSE, debug = FALSE)
}
\arguments{
\item{fun}{a function to be executed.}

\item{quiet}{a logical of whether display interactive message. Default FALSE
means display interactive message in execution.}

\item{debug}{a logical of whether execute fun in simulation mode or real
mode. Default FALSE means to execute function in real mode.}
}
\value{
a function which could be called in interactive mode.
}
\description{
A function normally needs be provided value for arguments to be called. We
could turn a normal function into a function with ability to be called
interactively, i.e. we could call it interactively by asking user to provide
value for each argument during execution.
}
\examples{
\dontrun{


# Turn print() into interactive function and call it interactively
inter_print <- interactive_fun(print)
inter_print()


# Turn print() into interactive function and call it interactively,
# execute it without displaying information
inter_print <- interactive_fun(print, quiet = TRUE)
inter_print()

# Turn print() into interactive function and call it interactively,
# execute it in debug mode
inter_print <- interactive_fun(print, debug = TRUE)
inter_print()
}
}
\seealso{
Other utils_cliui: 
\code{\link{help_fun}()},
\code{\link{interactive_call}()}
}
\concept{utils_cliui}
