% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-fields.R
\name{is_type_field}
\alias{is_type_field}
\title{Check whether fields are specified type}
\usage{
is_type_field(
  data,
  expect_type = c("numeric", "integer", "double", "character", "date", "factor",
    "list", "NA"),
  negate = FALSE,
  predicate_fun = NULL,
  ...
)
}
\arguments{
\item{data}{A vector of date/timestamps.}

\item{expect_type}{A character of type to test, e.g. "numeric", "integer",
"double", "character", "date", "factor", "list", "NA"}

\item{negate}{A logic flag of whether to negate expect_type of not,
Default FALSE means
not to negate the expect_type.}

\item{predicate_fun}{A function used as testing, return TRUE if matched.
Default NULL means not to use predicate_fun for testing.}

\item{...}{Params to predicate_fun.}
}
\value{
return a vectors of logical with same length of names of dataframe.
}
\description{
Predicate fields of dataframe are specified type or not.
}
\note{
numeric means integer or double; double don't include date.
}
\seealso{
Other utils_fields: 
\code{\link{expect_type_fields}()},
\code{\link{verify_fields}()}
}
\concept{utils_fields}
