% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-report.R
\name{knit_report}
\alias{knit_report}
\title{Custom Knit function for RStudio to build report}
\usage{
knit_report(
  input,
  report_params = "ask",
  output_format = NULL,
  output_sn = Sys.Date(),
  output_dir = "output",
  ...
)
}
\arguments{
\item{input}{The input file to be rendered as as report template, which is
required by customized knit function.}

\item{report_params}{A list of report params, by default NULL.
See params argument of \code{\link[rmarkdown:render]{rmarkdown::render()}} for details.}

\item{output_format}{The R Markdown output format to convert to.
See output_format argument of \code{\link[rmarkdown:render]{rmarkdown::render()}} for details.}

\item{output_sn}{Serial number or id for output file. Default NULL}

\item{output_dir}{Path of dir to save output file. Default "output" means to
save file in output in current dir. If NULL save in directory of template
file, otherwise in output_dir.}

\item{...}{Arguments passed to \code{\link[rmarkdown:render]{rmarkdown::render()}}.}
}
\description{
A wrapper of \code{\link[=build_report]{build_report()}} as a custom knit function for Knit button
in RStudio.
}
