% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock-db-gta.R, R/stock-db.R
\name{list_stock_tables.gta_db}
\alias{list_stock_tables.gta_db}
\alias{list_stock_tables}
\title{List all tables of stcok_db}
\usage{
\method{list_stock_tables}{gta_db}(stock_db)

list_stock_tables(stock_db)
}
\arguments{
\item{stock_db}{a stock database object to operate}
}
\value{
a vectors of characters of table names
}
\description{
Generic function to list all tables of stock_db
}
\section{Methods (by class)}{
\itemize{
\item \code{gta_db}: List names of tables in database of gta_db class
}}

\examples{
stock_db <- stock_db(gta_db, "GTA_SQLData")
open_stock_db(stock_db)
init_stock_db(stock_db)
list_stock_tables(stock_db)
}
