% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-factor-tools.R
\name{loadFactors}
\alias{loadFactors}
\title{Load factors Server Function}
\usage{
loadFactors(input, output, session, factors_info)
}
\arguments{
\item{input}{input from caller.}

\item{output}{output to caller.}

\item{session}{Session from caller.}

\item{factors_info}{A data.frame of factors.}
}
\description{
Load factors Server Function
}
