% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-factor-tools.R
\name{loadFactorsUI}
\alias{loadFactorsUI}
\title{Load factors UI}
\usage{
loadFactorsUI(id, factors_info)
}
\arguments{
\item{id}{An ID string of the UI.}

\item{factors_info}{A data.frame of factors.}
}
\description{
Load factors UI
}
