% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mode_value}
\alias{mode_value}
\title{get mode value of a vector}
\usage{
mode_value(x)
}
\arguments{
\item{x}{A vector of numbers, characters, factors}
}
\value{
A value with mode numbers
}
\description{
get mode value of a vector
}
\examples{
\dontrun{

# Get mode value of a vector of numeric
data <- c(1, 2, 2, 3, 3, 3, 4, 4, 5) #'
mode_value(data)

# Get model value of a vector of character
data <- c("A", "B", "B", "C", "C", "C", "D", "D", "E")
mode_value(data)

# Get model value of a vector of factor
data <- as.factor(c("A", "B", "B", "C", "C", "C", "D", "D", "E"))
mode_value(data)
}

}
