% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock-db.R, R/stock-db-gta.R
\docType{methods}
\name{name2code}
\alias{name2code}
\alias{name2code,gta_db-method}
\title{Translate name into code}
\usage{
name2code(x, name, ...)

\S4method{name2code}{gta_db}(x, name, ...)
}
\arguments{
\item{x}{a object containg code/name infomation}

\item{name}{a name or a vector of names to be translated}

\item{...}{other arguments to be provided to underlyling functions}
}
\value{
a code or a vector of codes
}
\description{
Generic function to translate name into code
}
\section{Methods (by class)}{
\itemize{
\item \code{gta_db}: Translate name into code in a database of gta_db class
}}

