% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor-normalize.R
\name{normalize}
\alias{normalize}
\title{Normalize a vector of data}
\usage{
normalize(x, clean_extremes_method = c("sigma", "mad"),
  standard_method = c("normal", "rank"), ...)
}
\arguments{
\item{x}{a vector of data.}

\item{clean_extremes_method}{method of cleaning extremes befre standardization
, i.e. "sigma","mad", by default "sigma".
see details: \code{\link{clean_extremes_sigma}}, \code{\link{clean_extremes_mad}}}

\item{standard_method}{method of standardizing data, i.e. "normal","rank",
by default "normal".
see details: \code{\link{standardize_normal_scale}}, \code{\link{standardize_rank_scale}}}

\item{...}{additional arguments to clean_extremes_method and standard_method.}
}
\value{
a vector of normalized data
}
\description{
Normalize a vector of data by cleaning extremes and standardizing.
}
