% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor-normalize.R
\name{normalize_factors}
\alias{normalize_factors}
\title{Normalize the factors dataset}
\usage{
normalize_factors(
  ds_factors,
  factors_list = NULL,
  group_by = NULL,
  clean_extremes_method = c("sigma", "mad"),
  standard_method = c("normal", "rank"),
  ...
)
}
\arguments{
\item{ds_factors}{A factors dataset.}

\item{factors_list}{A character vector of factors.}

\item{group_by}{A character vector of fields as group data for scaling, NULL
is default value which means no group setting.}

\item{clean_extremes_method}{Method of cleaning extremes before standardization
, i.e. "sigma","mad", by default "sigma".
see details: \code{\link{clean_extremes_sigma}}, \code{\link{clean_extremes_mad}}}

\item{standard_method}{method of standardizing data, i.e. "normal","rank",
by default "normal".
see details: \code{\link{standardize_normal_scale}}, \code{\link{standardize_rank_scale}}}

\item{...}{additional arguments to clean_extremes_method and standard_method.}
}
\value{
\preformatted{       a dataset of normalized factors
}
}
\description{
Normalize factors dataset by cleaning extremes and standardization.
}
