% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-factor-tools.R
\name{normalizeFactors}
\alias{normalizeFactors}
\title{Normalize factors Server Function}
\usage{
normalizeFactors(input, output, session, ds_factors)
}
\arguments{
\item{input}{input from caller.}

\item{output}{output to caller.}

\item{session}{Session from caller.}

\item{ds_factors}{A data.frame of factors.}
}
\description{
Normalize factors Server Function
}
