% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-factor-tools.R
\name{normalizeFactorsUI}
\alias{normalizeFactorsUI}
\title{Normalize factors UI}
\usage{
normalizeFactorsUI(id)
}
\arguments{
\item{id}{An ID string of the UI.}
}
\description{
Normalize factors UI
}
