% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicator-define-gta.R
\name{parse_indicator_vars.gta_db}
\alias{parse_indicator_vars.gta_db}
\title{database of gta_db class.}
\usage{
parse_indicator_vars.gta_db(stock_db, indicator_expr)
}
\description{
database of gta_db class.
}
