% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-mgt.R, R/data-mgt-gta.R
\docType{methods}
\name{process_files}
\alias{process_files}
\alias{process_files,gta_db-method}
\title{Process importing files in stock database}
\usage{
process_files(stock_db, data_source = get_datasource(stock_db),
  retry_log = NULL, log_file_prefix = "process_files_log",
  log_dir = "./log", ...)

\S4method{process_files}{gta_db}(stock_db,
  data_source = get_datasource(stock_db), retry_log = NULL,
  log_file_prefix = "process_files_log", log_dir = "./log", ...)
}
\arguments{
\item{stock_db}{A stock database object to operate.}

\item{data_source}{Data source info produced by get_datasource(stock_db)}

\item{retry_log}{Log file for re-processing tables with recording failure
in log file. If NULL, it will process all input_files in data_source, otherse
only process on these input_files with recording failure in log file.
By default NULL.}

\item{log_file_prefix}{A character of log file prefix to name log file.
  Log file is named as format of "log_file_prefix_XXXX_(current).csv"
Default is "update_db_log".}

\item{log_dir}{Path to save process log file. NULL means to use "./"
as log dir. Default "./log".}
}
\value{
NULL invisibly. Raise error if anything goes wrong.
}
\description{
Generic function to process importing files in stock database.
}
\details{
By combining infomation from data_source and re-try log file , it process
input_files of datasource in following steps:

\enumerate{
   \item Build input_files info from all input_files marked with process
   in datasource or some input_files with recoding failure in retry log file;
   \item Read input_files into original dataframe, process original
   dataframe by using process_fun specified by datasource, and save
   processed dataframe into files;
   \item Save update log, etc.
}

Process log is saved in log dir, like "process_files_log_XXXX(current).csv"
}
\section{Methods (by class)}{
\itemize{
\item \code{gta_db}: process importing files in a database of gta_db class
}}

\seealso{
Other data managment: \code{\link{clear_tables}},
  \code{\link{convert_import_file}},
  \code{\link{get_datasource}}, \code{\link{import_table}},
  \code{\link{read_import_file}},
  \code{\link{ttm_financial_report}},
  \code{\link{update_db}}, \code{\link{write_import_file}}
}
\concept{data managment}
