% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-mgt.R
\name{read_import_file}
\alias{read_import_file}
\title{Read a raw file for importing to stock db}
\usage{
read_import_file(input_file, input_type = c("csv", "txt"),
  input_dir = NULL, start_index = 2L, ignore_problems = TRUE,
  log_dir = "./log")
}
\arguments{
\item{input_file}{A name or a path of input data file.}

\item{input_type}{A charater of input file format, e.g. "txt", "csv".}

\item{input_dir}{A path of working dir of input file, if NULL, use dir of
input file as working dir, default NULL.}

\item{start_index}{A integer of start index of first line of actual records
in data file, by default 2L, which means first line is header and actual
data starts from second lines.}

\item{ignore_problems}{A logical to determine whether to ignore problems
when covert data, if TRUE, continue to import data but log problems into
log file, otherwise abort reading process.  By default TRUE.}

\item{log_dir}{Path of log dir for saving problem log file,
by default"./log", if the log path doesn't exsited, it  will be created.}
}
\value{
A dataframe of raw data for importing if succeed, otherwise NULL.
}
\description{
Read a raw file into a dataframe for processing or importing.
}
\seealso{
Other data managment: \code{\link{clear_tables}},
  \code{\link{convert_import_file}},
  \code{\link{get_datasource}}, \code{\link{import_table}},
  \code{\link{process_files}},
  \code{\link{ttm_financial_report}},
  \code{\link{update_db}}, \code{\link{write_import_file}}
}
\concept{data managment}
