% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicator-define.R
\name{related_indicator_defs}
\alias{related_indicator_defs}
\title{Get related indicator_defs of specified indicators}
\usage{
related_indicator_defs(ds_indicator_defs, indicator_codes)
}
\arguments{
\item{ds_indicator_defs}{A dataframe of indicators definition info.}

\item{indicator_codes}{A character vector of indicator codes.}
}
\value{
A dataframe of definitions of related indicators. Raise error if
anything goes wrong.
}
\description{
Get defs of related indicators of specified indicators by analyzing
dependency among indicators
}
\seealso{
Other indicator define functions: 
\code{\link{create_attribute_def_fun}()},
\code{\link{create_indicator_def_fun}()},
\code{\link{get_indicator_defs}()},
\code{\link{get_indicator_vars}()},
\code{\link{ind_attr_def_indcd}()},
\code{\link{ind_attr_def_trdstat}()},
\code{\link{parse_indicator_vars}()},
\code{\link{prioritize_indicator_defs}()}
}
\concept{indicator define functions}
