% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock-db-gta.R
\name{save_indicators_to_source.gta_db}
\alias{save_indicators_to_source.gta_db}
\title{specified source in a database of gta_db class}
\usage{
save_indicators_to_source.gta_db(stock_db, indicator_source, ts_indicators)
}
\description{
specified source in a database of gta_db class
}
