% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock-db.R, R/stock-db-gta.R
\name{save_indicators_to_source}
\alias{save_indicators_to_source}
\alias{save_indicators_to_source,gta_db-method}
\title{Save indicators to specified source in stock_db}
\usage{
save_indicators_to_source(stock_db, indicator_source, ts_indicators, ...)

\S4method{save_indicators_to_source}{gta_db}(stock_db, indicator_source, ts_indicators, ...)
}
\arguments{
\item{stock_db}{A stock database object to operate.}

\item{indicator_source}{A name of table or file(with extension)
in which indicators are stored. For file, formats of rds,csv are supported.}

\item{ts_indicators}{A dataframe of indicator timeseries}

\item{...}{Extra arguments to be passed to methods.}
}
\value{
NULL invisibly. Raise error if anything goes wrong.
}
\description{
Generic function to save indicator timeseries to specified source in stock_db.
}
\section{Functions}{
\itemize{
\item \code{save_indicators_to_source,gta_db-method}: save indicator timeseries to
specified source in a database of gta_db class
}}

\seealso{
Other stock_db generics: 
\code{\link{close_stock_db}()},
\code{\link{code2name}()},
\code{\link{dir_path_db}()},
\code{\link{fetch_table_dataset}()},
\code{\link{get_factors_info}()},
\code{\link{get_factors}()},
\code{\link{get_financial_report}()},
\code{\link{get_indicators_from_source}()},
\code{\link{get_indicators}()},
\code{\link{get_market_return}()},
\code{\link{get_profile}()},
\code{\link{get_riskfree_rate}()},
\code{\link{get_spt_stocks}()},
\code{\link{get_stock_dataset}()},
\code{\link{get_stock_industry}()},
\code{\link{get_stock_return}()},
\code{\link{get_table_dataset}()},
\code{\link{init_stock_db}()},
\code{\link{list_stock_tables}()},
\code{\link{name2code}()},
\code{\link{open_stock_db}()}
}
\concept{stock_db generics}
