% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-testthat-helpers.R
\name{skip_if_stock_db_not_ready}
\alias{skip_if_stock_db_not_ready}
\title{Skip tests if stock db is not ready}
\usage{
skip_if_stock_db_not_ready(dsn = "GTA_SQLData")
}
\arguments{
\item{dsn}{A string of DSN to connected. Default value "GTA_SQLData".}
}
\value{
No return value
}
\description{
This wrapper of \link[testthat:skip]{skip_if_not()} to skip tests
if stock db is not available for testing.
}
\details{
You could simulate a environment without stock db by setting environment
variable "NO_STOCK_DB" to "true".
}
\examples{
\dontrun{

# Skip tests when stock db is unavailable in the test file
skip_if_stock_db_not_ready()
test_that("tests related to stock db", {
  expect_equal(1, 3) #'
})

# Run Tests in a simulated environment without stock db
withr::with_envvar(
  new = c("NO_STOCK_DB" = "true"),
  devtools::test() # test package
  )
}

}
