% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-testthat-helpers.R
\name{skip_on_ci_for_os}
\alias{skip_on_ci_for_os}
\title{Skip on CI on specific OS}
\usage{
skip_on_ci_for_os(os)
}
\arguments{
\item{os}{Character vector of system names. Supported values are "windows",
"mac", "linux" and "solaris".}
}
\description{
skips on continuous integration systems like GitHub Actions, travis,
and appveyor (using the CI env var) on specific operating system.
}
\examples{
\dontrun{
 # Skip tests in CI testing on windows platform
   skip_on_ci_for_os("windows")
   test_that("tests is not good for CI on windows", {
     expect_equal(1, 3) #'
   })
}
}
