% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor-normalize.R
\name{standardize_normal_scale}
\alias{standardize_normal_scale}
\title{Standardize data vector by applying normal scale method}
\usage{
standardize_normal_scale(x)
}
\arguments{
\item{x}{a vector of data.}
}
\value{
\preformatted{ a vector of standardized data
}
}
\description{
Scale data into new vector of mean of 0 and stdev of 1
}
