% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor-normalize.R
\name{standardize_rank_scale}
\alias{standardize_rank_scale}
\title{Standardize data vector by applying rank scale method}
\usage{
standardize_rank_scale(x)
}
\arguments{
\item{x}{a vector of data.}
}
\value{
a vector of standardized data
}
\description{
Scale the ranked data by order
}
