% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock-db.R
\name{stock_db}
\alias{stock_db}
\title{subclass object factory of stock_db class}
\usage{
stock_db(stock_db_class, ...)
}
\arguments{
\item{stock_db_class}{specific creator of stock database class,
e.g. gta_db}

\item{...}{addition params used by specific class of stock database(e.g. a dsn string
"GTA_SQLData")}
}
\value{
a object of stock db
}
\description{
factory for creating for subclass object of stock_db class
}
\examples{
stock_db <- stock_db(gta_db, "GTA_SQLData")

}
