% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock-db.R
\name{stock_db_operation}
\alias{stock_db_operation}
\title{Generic function for stock_db
Generic function of operating stock database}
\description{
A group generic function to deal with a stock_db object
}
\examples{
\dontrun{
library(zstmodelr)
stock_db <- stock_db(gta_db, "GTA_SQLData")
open_stock_db(stock_db)
close_stock_db(stock_db)
}
}
