% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock-db-gta.R
\name{stock_field_list.gta_db}
\alias{stock_field_list.gta_db}
\title{Create a stock_filed_list for a database of gta_db class}
\usage{
stock_field_list.gta_db(stock_db)
}
\description{
Create a stock_filed_list for a database of gta_db class
}
