% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock-db-gta.R, R/stock-db.R
\name{stock_field_list.gta_db}
\alias{stock_field_list.gta_db}
\alias{stock_field_list}
\title{Class creator of stock_field class}
\usage{
\method{stock_field_list}{gta_db}(stock_db)

stock_field_list(stock_db)
}
\arguments{
\item{stock_db}{a stock database object to provide stock_field info}
}
\value{
a object of stock_filed class
}
\description{
Generic stock_field class creator
}
\section{Methods (by class)}{
\itemize{
\item \code{gta_db}: create a stock_filed_list for a database of gta_db class
}}

\examples{

}
