% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock-db-gta.R, R/stock-db.R
\name{stock_name_list.gta_db}
\alias{stock_name_list.gta_db}
\alias{stock_name_list}
\title{Class creator of stock_name_list class}
\usage{
\method{stock_name_list}{gta_db}(stock_db)

stock_name_list(stock_db)
}
\arguments{
\item{stock_db}{a stock database object to provide stock_name info}
}
\value{
a object of stock_name class
}
\description{
generic stock_name_list class creator
}
\section{Methods (by class)}{
\itemize{
\item \code{gta_db}: create a stock_name_list for a database of gta_db class
}}

