% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{suppress_warnings}
\alias{suppress_warnings}
\title{Suppress specific warnings}
\usage{
suppress_warnings(expr, warn_pattern = NA)
}
\arguments{
\item{expr}{A expression to evaluate.}

\item{warn_pattern}{A character pattern as a regular expression. Default NA
means suppressing all warnings.}
}
\value{
The warning message as character string, invisibly.
}
\description{
Unlike \code{\link[base:suppressWarnings]{base::suppressWarnings()}} which suppresses all warnings, it can
suppress some specific warnings and output remain warnings.
}
\examples{
suppress_warnings(
  {
    sqrt(-1)
    warning("ooops", call. = FALSE)
  },
  warn_pattern = "NaN"
)
suppress_warnings(
  {
    sqrt(-1)
    warning("ooops", call. = FALSE)
  },
  warn_pattern = "o"
)
}
