% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-timeseries.R
\name{ts_lag}
\alias{ts_lag}
\alias{ts_lag.tbl_df}
\alias{ts_lag.timeSeries}
\title{Compute a lagged version of timeseries}
\usage{
ts_lag(ts_dataset, k = 1, trim = TRUE, ..., date_index_field = c("date"),
  by_group = NULL)

\method{ts_lag}{tbl_df}(ts_dataset, k = 1, trim = TRUE, ...,
  date_index_field = c("date"), by_group = NULL)

\method{ts_lag}{timeSeries}(ts_dataset, k = 1, trim = TRUE, ...)
}
\arguments{
\item{ts_dataset}{a timeseries of tibble/timeSeries.}

\item{k}{an integer value. The number of lags (in units of observations).
By default 1.}

\item{trim}{a logical value. By default TRUE, the first missing observation in the return series
will be removed.}

\item{...}{argments passed to other methods}

\item{date_index_field}{the name of date index field of ts_df for resample,
default 'date', Column must be date-like.
Only be used for tibble dataset.}

\item{by_group}{a character vector of fields as group data for asfreq.
Only be used for tibble dataset.}
}
\value{
a lagged timeseres
}
\description{
Compute a lagged version of timeseries
}
\section{Methods (by class)}{
\itemize{
\item \code{tbl_df}: Compute a lagged version of timeseries for tibble dataset

\item \code{timeSeries}: Compute a lagged version of timeseries for timeSeries dataset
}}

