% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datamgt.R, R/datamgt-gta.R
\docType{methods}
\name{update_db}
\alias{update_db}
\alias{update_db,gta_db-method}
\title{Update tables in stock database with new data}
\usage{
update_db(stock_db, data_source = get_datasource(stock_db),
  retry_log = NULL, log_dir = "./log", ...)

\S4method{update_db}{gta_db}(stock_db, data_source = get_datasource(stock_db),
  retry_log = NULL, log_dir = "./log", ...)
}
\arguments{
\item{stock_db}{A stock database object to operate.}

\item{data_source}{Data source info produced by get_datasource(stock_db)}

\item{retry_log}{Log file for re-importing tables with recording failure
in log file. If NULL, it will upate all talbles in data_source, otherse
only update on these tables with recording failure in log file.
By default NULL.}

\item{log_dir}{Path to save updating log file. If NULL, don't save log file
by default "./log".}
}
\value{
TRUE if success, else FALSE.
}
\description{
Generic function to update tables in stock database by importing new data
}
\details{
By combining infomation from data_source and re-try log file , it update all
tables from raw data of datasource in following steps:

\enumerate{
   \item build target tables info from all tables in datasource or some
   tables with recoding failure in retry log file.
   \item update target tables by importing data from datasoure.
   \item save update log, etc.
}

Update log is saved in log dir, like "update_log_XXXX(current).csv"
}
\section{Methods (by class)}{
\itemize{
\item \code{gta_db}: update tables in a database of gta_db class
}}

