% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-expr.R
\name{utils_expr}
\alias{utils_expr}
\alias{create_expr}
\alias{find_syms}
\title{Utils functions of expr}
\usage{
create_expr(expr)

find_syms(expr, pattern = NULL, ...)
}
\arguments{
\item{expr}{A expr or list of exprs or string to parse as expr.}

\item{pattern}{a character of regular rule for matching.}
}
\description{
Utility functions to handle expr basing on rlang, which is used to facliate
defining indicator.
}
\section{Functions}{
\itemize{
\item \code{create_expr}: create a expr for evaluation from expr,
list of exprs, or strings.

\item \code{find_syms}: Find symbols in expr by matching pattern.
}}

