% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-profile.R
\name{utils_profile}
\alias{utils_profile}
\alias{get_profile_path}
\alias{profile_get_varible_setting}
\alias{profile_get_factors}
\alias{profile_get_datasource_files}
\alias{profile_get_indicators}
\alias{profile_get_customized_indicators}
\title{Utils functions of profile}
\usage{
get_profile_path(profile_name, profile_dir = .PROFILE_DIR)

profile_get_varible_setting(profile_path, variable)

profile_get_factors(profile_path, factor_codes = NULL,
  factor_groups = NULL)

profile_get_datasource_files(profile_path)

profile_get_indicators(profile_path, indicator_codes = NULL)

profile_get_customized_indicators(profile_path)
}
\arguments{
\item{profile_name}{A character of profile name, e.g. .GTA_PROFILE_FILE
or "gta_profile.xlsx".}

\item{profile_dir}{A character name of dir where profile is located.
Default .PROFILE_DIR is etc in package directory.}

\item{profile_path}{path of profile to get data from.}

\item{variable}{A character of name of variable to get.}

\item{factor_codes}{A character vector of factor codes to fetch.
Default NULL means to fetch all factors.}

\item{factor_groups}{A character vector of factor groups to fetch.
Default NULL means to fetch all factor groups.}

\item{indicator_codes}{A character vector of indicator codes to fetch.
Default NULL means to fetch all indicators.}
}
\value{
\strong{get_profile_path}: A path of profile if succeed, otherwise "".

\strong{others}: A dataframe of setting if succeed, othwise NULL.
}
\description{
Utility functions to handle profile, which contains configration
information of stock database.
}
\section{Functions}{
\itemize{
\item \code{get_profile_path}: get full path of profile.

\item \code{profile_get_varible_setting}: get table name value from profile of database.

\item \code{profile_get_factors}: get table name value from profile of database.

\item \code{profile_get_datasource_files}: get data source for importing raw data.

\item \code{profile_get_indicators}: get info of indicators from profile of database.

\item \code{profile_get_customized_indicators}: get info of customized indicators from profile
of database.
}}

