% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor-composite.R
\name{weight_formula}
\alias{weight_formula}
\title{Build aggregating formula of composite factor}
\usage{
weight_formula(factors_weight, new_factor_name = "composite_factor")
}
\arguments{
\item{factors_weight}{a list of factor weight with columns of factor_name and
factor_weight.}

\item{new_factor_name}{a name of composite factor, by default "composite_factor".}
}
\value{
\preformatted{       a formula of compoiste factor,
}

like \code{composite_factor ~ 0.33*PE + 0.33*PB + 0.33*PS}.
}
\description{
Aggregating formula is built from weight of similar or correlated factors
Composite factor is built by aggregating formula coming from weight of
similar or correlated factors, which is determined by various methods, such as ....
}
