% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-withr-helpers.R
\name{with_parallel}
\alias{with_parallel}
\alias{local_parallel}
\title{Turn on/off Parallel process temporarily}
\usage{
with_parallel(new, code)

local_parallel(new = list(), .local_envir = parent.frame())
}
\arguments{
\item{new}{\verb{[character(1)]} new parallel state: "ON" or "OFF".}
}
\description{
Temporarily turn on/off parallel process.
}
\examples{
\dontrun{

with_parallel(new = "ON", {
  sprintf(
    "Parallel prcoess is \%s.",
    ifelse(parallel_is_on(), "ON", "OFF")
  )
})

with_parallel(new = "OFF", {
  sprintf(
    "Parallel prcoess is \%s.",
    ifelse(parallel_is_on(), "ON", "OFF")
  )
})

local_parallel("ON")
sprintf(
  "Parallel prcoess is \%s.",
  ifelse(parallel_is_on(), "ON", "OFF")
)
local_parallel("OFF")
sprintf(
  "Parallel prcoess is \%s.",
  ifelse(parallel_is_on(), "ON", "OFF")
)
}
}
