% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-mgt.R
\name{write_import_file}
\alias{write_import_file}
\title{Write a raw file for importing to stock db}
\usage{
write_import_file(ds_output, output_file, output_type = c("csv", "txt"),
  output_dir = NULL)
}
\arguments{
\item{ds_output}{A name or a path of input data file.}

\item{output_file}{A name or a path of input data file.}

\item{output_type}{A character of output file format, e.g. "txt", "csv".}

\item{output_dir}{A path of working dir of output file, if NULL, use dir
of output file as working dir, by default NULL,}
}
\value{
NULL invisibly. Raise error if anything goes wrong.
}
\description{
Write a dataframe as import file. If output file existed, original file will
be backed up.
}
\seealso{
Other data managment: \code{\link{clear_tables}},
  \code{\link{convert_import_file}},
  \code{\link{get_datasource}}, \code{\link{import_table}},
  \code{\link{process_files}},
  \code{\link{read_import_file}},
  \code{\link{ttm_financial_report}},
  \code{\link{update_db}}
}
\concept{data managment}
