% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zstmodelr.R
\docType{package}
\name{zstmodelr}
\alias{zstmodelr}
\alias{zstmodelr-package}
\title{ZStockModels: A package for computating the notorious bar statistic.}
\description{
The foo package provides three categories of important functions:
foo, bar and baz.
}
\section{tools functions}{

The add functions ...
}

