% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-report.R
\name{build_report}
\alias{build_report}
\title{Build report from template}
\usage{
build_report(
  report_template,
  report_params = NULL,
  output_format = c("html_document", "pdf_document", "word_document", "rticles::ctex"),
  output_sn = NULL,
  output_dir = "output",
  quiet = TRUE,
  ...
)
}
\arguments{
\item{report_template}{A rmd file as report template.}

\item{report_params}{A list of report params, by default NULL.}

\item{output_format}{Output file format i.e. "html_document","pdf_document",
"word_document", "rticles::ctex". Default "html_document".}

\item{output_sn}{Serial number or id for output file. Default NULL}

\item{output_dir}{Path of dir to save output file. Default "output" means to
save file in output in current dir. If NULL save in directory of template
file, otherwise in output_dir}

\item{quiet}{A logic flag of whether to suppress printing of the pandoc
command line.}

\item{...}{Arguments passed to rmarkdown::render.}
}
\value{
No return.
}
\description{
Use data rank as new data
}
\concept{utils_report}
