% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor-test.R
\name{build_sort_portfolios}
\alias{build_sort_portfolios}
\title{Build sort portfoilos basing on group list of stocks}
\usage{
build_sort_portfolios(
  stocks_list,
  factor_value_list,
  stocks_weight_list = NULL,
  ngroup = 5
)
}
\arguments{
\item{stocks_list}{A list of stkcds of stocks.}

\item{factor_value_list}{A list of factor value of corresponding stocks.}

\item{stocks_weight_list}{A list of stocks weight or NULL for equal weight of 1,
by default NULL.}

\item{ngroup}{Numbers of groups to cut stocks, by default 5.}
}
\value{
\preformatted{            A tibble dataset of portfolio_group, stkcd, weight,
                    factor_value.
}
}
\description{
use stocks list and factor value list to build sort portfoilos by sorting
factor value and cutting in n groups.
}
