% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-mgt.R, R/data-mgt-gta.R
\docType{methods}
\name{clear_tables}
\alias{clear_tables}
\alias{clear_tables,gta_db-method}
\title{Clear data in all tables in stock database}
\usage{
clear_tables(stock_db, ...)

\S4method{clear_tables}{gta_db}(stock_db, ...)
}
\arguments{
\item{stock_db}{A stock database object to operate.}
}
\value{
NULL invisibly. Raise error if anything goes wrong.
}
\description{
Generic function to Clear existed data in all tables in stock_db.
}
\section{Methods (by class)}{
\itemize{
\item \code{gta_db}: clear data in all tables in a database of gta_db class
}}

\seealso{
Other data managment: \code{\link{convert_import_file}},
  \code{\link{get_datasource}}, \code{\link{import_table}},
  \code{\link{process_files}},
  \code{\link{read_import_file}},
  \code{\link{ttm_financial_report}},
  \code{\link{update_db}}, \code{\link{write_import_file}}
}
\concept{data managment}
