% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock-db.R, R/stock-db-gta.R
\docType{methods}
\name{code2name}
\alias{code2name}
\alias{code2name,gta_db-method}
\title{Translate code into name}
\usage{
code2name(x, code, ...)

\S4method{code2name}{gta_db}(x, code, ...)
}
\arguments{
\item{x}{A object containg code/name infomation}

\item{code}{A code or a vector of codes to be translated}

\item{...}{Other arguments to be provided to underlyling functions}
}
\value{
A name or a vector of names
}
\description{
Generic function to translate code into name
}
\section{Methods (by class)}{
\itemize{
\item \code{gta_db}: Translate code into name in a database of gta_db class
}}

\seealso{
Other stock_db generics: \code{\link{close_stock_db}},
  \code{\link{fetch_table_dataset}},
  \code{\link{get_factor_indicator}},
  \code{\link{get_factors_info}},
  \code{\link{get_market_return}},
  \code{\link{get_stock_dataset}},
  \code{\link{get_stock_return}},
  \code{\link{get_table_dataset}},
  \code{\link{init_stock_db}},
  \code{\link{list_stock_tables}}, \code{\link{name2code}},
  \code{\link{open_stock_db}}
}
