% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock-db.R, R/code-name-list.R,
%   R/stock-db-gta.R
\name{code2name}
\alias{code2name}
\alias{code2name,code_name_list-method}
\alias{code2name,gta_db-method}
\title{Translate code into name}
\usage{
code2name(x, code, ...)

\S4method{code2name}{code_name_list}(x, code, ...)

\S4method{code2name}{gta_db}(x, code, ...)
}
\arguments{
\item{x}{A object containing code/name information.}

\item{code}{A code or a vector of codes to be translated.}

\item{...}{Extra arguments to be passed to methods.}
}
\value{
\preformatted{A name or a vector of names.
}
}
\description{
Generic function to translate code into name
}
\section{Functions}{
\itemize{
\item \code{code2name,code_name_list-method}: Translate code into name in a object of code_name_list

\item \code{code2name,gta_db-method}: Translate code into name in a database of gta_db class
}}

\seealso{
Other stock_db generics: 
\code{\link{close_stock_db}()},
\code{\link{dir_path_db}()},
\code{\link{fetch_table_dataset}()},
\code{\link{get_factors_info}()},
\code{\link{get_factors}()},
\code{\link{get_financial_report}()},
\code{\link{get_indicators_from_source}()},
\code{\link{get_indicators}()},
\code{\link{get_market_return}()},
\code{\link{get_profile}()},
\code{\link{get_riskfree_rate}()},
\code{\link{get_spt_stocks}()},
\code{\link{get_stock_dataset}()},
\code{\link{get_stock_industry}()},
\code{\link{get_stock_return}()},
\code{\link{get_table_dataset}()},
\code{\link{init_stock_db}()},
\code{\link{list_stock_tables}()},
\code{\link{name2code}()},
\code{\link{open_stock_db}()},
\code{\link{save_indicators_to_source}()}
}
\concept{stock_db generics}
