% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock-db-gta.R, R/stock-db.R
\name{code2name.gta_db}
\alias{code2name.gta_db}
\alias{code2name}
\alias{code2name.code_name_list}
\title{Translate code into name}
\usage{
\method{code2name}{gta_db}(stock_db, code, type = c("stock", "field",
  "industry"))

code2name(x, code, ...)

\method{code2name}{code_name_list}(x, code)
}
\arguments{
\item{code}{a code or a vector of codes to be translated}

\item{x}{a object containg code/name infomation}

\item{...}{other params to be provided to underlyling functions}
}
\value{
a name or a vector of names
}
\description{
Generic function to translate code into name
}
\section{Methods (by class)}{
\itemize{
\item \code{gta_db}: Translate code into name in a database of gta_db class

\item \code{code_name_list}: Translate code into name in a object of code_name_list
}}

\examples{

}
