% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-factor.R
\name{factor_test_uniregress}
\alias{factor_test_uniregress}
\title{Creator of factor_test_unigress class}
\usage{
factor_test_uniregress(ds_test, regress_method = c("cross_section",
  "pooling"), regress_fun, ..., output_type = c("summary", "raw"),
  factor_field = "factor_name", date_field = "date")
}
\arguments{
\item{ds_test}{a timeseries dataset with descriptors of factors for test}

\item{regress_method}{method of partitioning data for regression, i.g.
"cross_section", "pooling", by default "cross_section"}

\item{regress_fun}{a function to conduct regress}

\item{...}{argments passed to regress_fun}

\item{output_type}{type of output data, i.g."summary", "raw", if "raw",
raw data will be append to output object for dignosis}

\item{factor_field}{the name of factor field of ds_test, by default "factor_name"}

\item{date_field}{the name of date field of ds_test, by default "date",
Column must be date-like.}
}
\value{
a object of factor_test_uniregress class
}
\description{
Conduct univarate reggression test for descriptors of factors and build object
of factor_test_uniregress class as output
}
