% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-zscore.R
\name{factors_zscore}
\alias{factors_zscore}
\title{Compute z-scores of facotrs}
\usage{
factors_zscore(ds_factors, aggregate_formula = NULL, group_by = NULL)
}
\arguments{
\item{ds_factors}{a factors dataset}

\item{aggregate_formula}{a formula to aggreate zscores of all factors,
e.g. ROCE + P/B, default is NULL which means no need to
compute aggregate zscore, but only compute zscore for all
numberic factor fields}

\item{group_by}{a character vector of fields as group data for scaling}
}
\value{
a z-score datasets of factors
}
\description{
Compute z-scores of factors by scaling all factors according the aggreate
formula and groups of fields.
The z-scores are in the range of [-3, 3] by
replacing the values of outliers beyond(-3/3) with -3 or 3
}
