% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-mgt.R, R/data-mgt-gta.R
\docType{methods}
\name{get_datasource}
\alias{get_datasource}
\alias{get_datasource,gta_db-method}
\title{Get data source info for importing raw data}
\usage{
get_datasource(stock_db, ...)

\S4method{get_datasource}{gta_db}(stock_db, ...)
}
\arguments{
\item{stock_db}{A stock database object to operate.}
}
\value{
A dataframe of info about importing raw data if succeed,
otherwise NULL.
}
\description{
Generic function to get data source info for importing raw data into stock_db.
}
\section{Methods (by class)}{
\itemize{
\item \code{gta_db}: get data source for importing raw data into
database of gta_db class
}}

\seealso{
Other data managment: \code{\link{clear_tables}},
  \code{\link{convert_import_file}},
  \code{\link{import_table}}, \code{\link{process_files}},
  \code{\link{read_import_file}},
  \code{\link{ttm_financial_report}},
  \code{\link{update_db}}, \code{\link{write_import_file}}
}
\concept{data managment}
