% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datamgt.R
\name{get_datasource}
\alias{get_datasource}
\title{Get data source info for importing raw data}
\usage{
get_datasource(stock_db, ...)
}
\arguments{
\item{stock_db}{A stock database object to operate.}
}
\value{
A dataframe of info about importing raw data if successfully,
otherwise NULL.
}
\description{
Generic function to get data source info for importing raw data into stock_db.
}
\seealso{
Other data managment functions: \code{\link{clear_tables}},
  \code{\link{import_table}}
}
