% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock-db.R, R/stock-db-gta.R
\docType{methods}
\name{get_factors_info}
\alias{get_factors_info}
\alias{get_factors_info,gta_db-method}
\title{Get factors Info from stock_db}
\usage{
get_factors_info(stock_db, factor_groups = NULL, ...)

\S4method{get_factors_info}{gta_db}(stock_db, factor_groups = NULL, ...)
}
\arguments{
\item{stock_db}{A stock database object to operate.}

\item{factor_groups}{A character vector of factor groups, if NULL will
return all factors, otherwise return factors of matched fator groups,
by default NULL.}
}
\value{
A dataframe of factors of matched factor groups if successfully,
otherwise NULL.
}
\description{
Generic function to get factors info from stock_db.
}
\section{Methods (by class)}{
\itemize{
\item \code{gta_db}: get factors info from a database of gta_db class
}}

\seealso{
Other stock_db generics: \code{\link{close_stock_db}},
  \code{\link{code2name}},
  \code{\link{fetch_table_dataset}},
  \code{\link{get_factor_indicator}},
  \code{\link{get_market_return}},
  \code{\link{get_stock_dataset}},
  \code{\link{get_stock_return}},
  \code{\link{get_table_dataset}},
  \code{\link{init_stock_db}},
  \code{\link{list_stock_tables}}, \code{\link{name2code}},
  \code{\link{open_stock_db}}
}
