% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicator-define.R, R/indicator-define-gta.R
\name{get_indicator_defs}
\alias{get_indicator_defs}
\alias{get_indicator_defs,gta_db-method}
\title{Get definitions of customized indicators from stock_db}
\usage{
get_indicator_defs(stock_db, ...)

\S4method{get_indicator_defs}{gta_db}(stock_db, ...)
}
\arguments{
\item{stock_db}{A stock database object to operate.}

\item{...}{Extra arguments to be passed to methods.}
}
\value{
A dataframe of definitions of customized indicators if succeed,
otherwise NULL.
}
\description{
Generic function to get definitions of customized indicators from stock_db.
}
\section{Functions}{
\itemize{
\item \code{get_indicator_defs,gta_db-method}: get indicator defs from database
of gta_db class.
}}

\seealso{
Other indicator define functions: 
\code{\link{create_attribute_def_fun}()},
\code{\link{create_indicator_def_fun}()},
\code{\link{get_indicator_vars}()},
\code{\link{ind_attr_def_indcd}()},
\code{\link{ind_attr_def_trdstat}()},
\code{\link{parse_indicator_vars}()},
\code{\link{prioritize_indicator_defs}()},
\code{\link{related_indicator_defs}()}
}
\concept{indicator define functions}
