% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock-db.R, R/stock-db-gta.R
\name{get_market_return}
\alias{get_market_return}
\alias{get_market_return,gta_db-method}
\title{Get market return timeseries from stock_db}
\usage{
get_market_return(
  stock_db,
  period_type = c("day", "month", "year"),
  period_date = c("start", "end"),
  output_type = c("timeSeries", "tibble"),
  ...
)

\S4method{get_market_return}{gta_db}(
  stock_db,
  period_type = c("day", "month", "year"),
  period_date = c("start", "end"),
  output_type = c("timeSeries", "tibble"),
  ...
)
}
\arguments{
\item{stock_db}{A stock database object to operate.}

\item{period_type}{Date period for time series, e.g. "day", "month",
"year", by default value is  "day".}

\item{period_date}{Choose start/end date of period as the date for
observation}

\item{output_type}{Format of output result, e.g "timeSeries", "tibble".}

\item{...}{Extra arguments to be passed to methods.}
}
\value{
A timeseries of market return.
}
\description{
Generic function to get stock return timeseries from stock_db.
}
\section{Functions}{
\itemize{
\item \code{get_market_return,gta_db-method}: get market return timeseries from a database of gta_db class
}}

\seealso{
Other stock_db generics: 
\code{\link{close_stock_db}()},
\code{\link{code2name}()},
\code{\link{dir_path_db}()},
\code{\link{fetch_table_dataset}()},
\code{\link{get_factors_info}()},
\code{\link{get_factors}()},
\code{\link{get_financial_report}()},
\code{\link{get_indicators_from_source}()},
\code{\link{get_indicators}()},
\code{\link{get_profile}()},
\code{\link{get_riskfree_rate}()},
\code{\link{get_spt_stocks}()},
\code{\link{get_stock_dataset}()},
\code{\link{get_stock_industry}()},
\code{\link{get_stock_return}()},
\code{\link{get_table_dataset}()},
\code{\link{init_stock_db}()},
\code{\link{list_stock_tables}()},
\code{\link{name2code}()},
\code{\link{open_stock_db}()},
\code{\link{save_indicators_to_source}()}
}
\concept{stock_db generics}
