% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock-db-gta.R, R/stock-db.R
\name{get_market_return.gta_db}
\alias{get_market_return.gta_db}
\alias{get_market_return}
\title{Generic function to get stock return timeseries from stock_db}
\usage{
\method{get_market_return}{gta_db}(stock_db, period_type = c("daily",
  "weekly", "monthly", "annual"))

get_market_return(stock_db, period_type = c("daily", "weekly", "monthly",
  "annual"))
}
\arguments{
\item{stock_db}{a stock database object to operate}

\item{period_type}{date peroid for time series, e.g. "daily", "weekly",
"monthly", "annual", default value is  "daily"}
}
\value{
A timeseries of market return
}
\description{
Generic function to get stock return timeseries from stock_db
}
\section{Methods (by class)}{
\itemize{
\item \code{gta_db}: get market return timeseries from a database of gta_db class
}}

\examples{

}
