% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock-db.R, R/stock-db-gta.R
\docType{methods}
\name{get_stock_dataset}
\alias{get_stock_dataset}
\alias{get_stock_dataset,gta_db-method}
\title{Get a dataset of a list of stock_cd from a table in stock_db}
\usage{
get_stock_dataset(stock_db, table_name, stock_cd_list = NULL, ...)

\S4method{get_stock_dataset}{gta_db}(stock_db, table_name,
  stock_cd_list = NULL, ...)
}
\arguments{
\item{stock_db}{A stock database object to operate.}

\item{table_name}{Name of target table.}

\item{stock_cd_list}{A list of stock cd, default value of NULL means.
all stock data will be returned.}
}
\value{
A data frame on succeed otherwise NULL.
}
\description{
Generic function to get a dataset of a list of stock_cd from table in stock.
}
\section{Methods (by class)}{
\itemize{
\item \code{gta_db}: get a dataset of a list of stock_cd from table
in a database of gta_db class
}}

\examples{
ds_trd_dalyr.df <- get_stock_dataset(stock_db,
                            table_name = "TRD_Dalyr_日个股回报率",
                            stock_cd_list = c("600066", "000550"))
}
\seealso{
Other stock_db generics: \code{\link{close_stock_db}},
  \code{\link{code2name}}, \code{\link{dir_path_db}},
  \code{\link{fetch_table_dataset}},
  \code{\link{get_factors_info}},
  \code{\link{get_factors}},
  \code{\link{get_indicators_from_source}},
  \code{\link{get_indicators}},
  \code{\link{get_market_return}},
  \code{\link{get_profile}},
  \code{\link{get_stock_return}},
  \code{\link{get_table_dataset}},
  \code{\link{init_stock_db}},
  \code{\link{list_stock_tables}}, \code{\link{name2code}},
  \code{\link{open_stock_db}},
  \code{\link{save_indicators_to_source}}
}
\concept{stock_db generics}
