% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock-db-gta.R, R/stock-db.R
\name{get_stock_return.gta_db}
\alias{get_stock_return.gta_db}
\alias{get_stock_return}
\title{Get stock return timeseries from stock_db}
\usage{
\method{get_stock_return}{gta_db}(stock_db, stock_cd_list = NULL,
  period_type = c("daily", "weekly", "monthly", "annual"), handleNA = c("r",
  "s", "z", "ir", "iz", "ie"), interpNA_method = c("before", "linear",
  "after"))

get_stock_return(stock_db, stock_cd_list = NULL, period_type = c("daily",
  "weekly", "monthly", "annual"), handleNA = c("r", "s", "z", "ir", "iz",
  "ie"), interpNA_method = c("before", "linear", "after"))
}
\arguments{
\item{stock_db}{a stock database object to operate}

\item{stock_cd_list}{a list of stock cd, default value of NULL means
all stock data will be returned}

\item{period_type}{date peroid for time series, e.g. "daily", "weekly",
"monthly", "annual", default value is "daily"}

\item{handleNA}{Specifies the method how to handle NAs. One of the applied vector strings:
method="s" na.rm = FALSE, skip, i.e. do nothing, method="r" remove NAs,
method="z" substitute NAs by zeros, method="ir" interpolate NAs and remove
NAs at the beginning and end of the series, method="iz" interpolate NAs and
substitute NAs at the beginning and end of the series, method="ie" interpolate
NAs and extrapolate NAs at the beginning and end of the series#'}

\item{interpNA_method}{Specifies the method how to interpolate NA the matrix column by column.
One of the applied vector strings: method="linear", method="before"
or method="after". For the interpolation the function approx is used.}
}
\value{
A timeseries of stock return
}
\description{
Generic function to get stock return timeseries from stock_db
}
\section{Methods (by class)}{
\itemize{
\item \code{gta_db}: get stock return timeseries from a database of gta_db class
}}

\examples{

}
