% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock-db.R, R/stock-db-gta.R
\docType{methods}
\name{get_stock_return}
\alias{get_stock_return}
\alias{get_stock_return,gta_db-method}
\alias{get_factor_indicator,gta_db-method}
\title{Get stock return timeseries from stock_db}
\usage{
get_stock_return(stock_db, stock_cd_list = NULL, period_type = c("day",
  "month", "year"), period_date = c("start", "end"),
  output_type = c("timeSeries", "tibble"), ...)

\S4method{get_stock_return}{gta_db}(stock_db, stock_cd_list = NULL,
  period_type = c("day", "month", "year"), period_date = c("start", "end"),
  output_type = c("timeSeries", "tibble"), ...)

\S4method{get_factor_indicator}{gta_db}(stock_db, factor_list, ...)
}
\arguments{
\item{stock_db}{A stock database object to operate.}

\item{stock_cd_list}{A list of stock cd, default value of NULL means
all stock data will return.}

\item{period_type}{Date peroid for time series, e.g. "day", "month",
"year", by default "day".}

\item{period_date}{Choose start/end date of period as the date for
observation}

\item{output_type}{Format of output result, e.g "timeSeries", "tibble".}
}
\value{
A timeseries of stock return.
}
\description{
Generic function to get stock return timeseries from stock_db.
}
\section{Methods (by class)}{
\itemize{
\item \code{gta_db}: get stock return timeseries from a database of gta_db class

\item \code{gta_db}: get stock return timeseries from a database of gta_db class
}}

\seealso{
Other stock_db generics: \code{\link{close_stock_db}},
  \code{\link{code2name}},
  \code{\link{fetch_table_dataset}},
  \code{\link{get_factor_indicator}},
  \code{\link{get_factors_info}},
  \code{\link{get_market_return}},
  \code{\link{get_stock_dataset}},
  \code{\link{get_table_dataset}},
  \code{\link{init_stock_db}},
  \code{\link{list_stock_tables}}, \code{\link{name2code}},
  \code{\link{open_stock_db}}
}
