% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock-db-gta.R, R/stock-db.R
\name{get_stock_return.gta_db}
\alias{get_stock_return.gta_db}
\alias{get_factor_indicator.gta_db}
\alias{get_stock_return}
\title{Get stock return timeseries from stock_db}
\usage{
\method{get_stock_return}{gta_db}(stock_db, stock_cd_list = NULL,
  period_type = c("daily", "weekly", "monthly", "annual"),
  return_type = c("simple", "compound"), use_stock_name = TRUE,
  cumulated = FALSE, output_type = c("timeSeries", "tibble"))

\method{get_factor_indicator}{gta_db}(stock_db, factor_list)

get_stock_return(stock_db, stock_cd_list = NULL, period_type = c("daily",
  "weekly", "monthly", "annual"), return_type = c("simple", "compound"),
  use_stock_name = TRUE, cumulated = FALSE, output_type = c("timeSeries",
  "tibble"))
}
\arguments{
\item{stock_db}{a stock database object to operate.}

\item{stock_cd_list}{a list of stock cd, default value of NULL means
all stock data will be returned.}

\item{period_type}{date peroid for time series, e.g. "daily", "weekly",
"monthly", "annual", default value is "daily".}

\item{return_type}{a character string naming the method how the returns were computed.}

\item{use_stock_name}{use stock name as field name instead of stock cd, default is TRUE.}

\item{cumulated}{calculate cumulated return, default value is FALSE.}

\item{output_type}{format of output result, e.g "timeSeries", "tibble".}
}
\value{
A timeseries of stock return
}
\description{
Generic function to get stock return timeseries from stock_db
}
\section{Methods (by class)}{
\itemize{
\item \code{gta_db}: get stock return timeseries from a database of gta_db class

\item \code{gta_db}: get stock return timeseries from a database of gta_db class
}}

