% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock-db.R, R/stock-db-gta.R
\docType{methods}
\name{get_table_dataset}
\alias{get_table_dataset}
\alias{get_table_dataset,gta_db-method}
\title{Get adataset from a table in stock_db}
\usage{
get_table_dataset(stock_db, table_name, ...)

\S4method{get_table_dataset}{gta_db}(stock_db, table_name, ...)
}
\arguments{
\item{stock_db}{a stock database object to operate.}

\item{table_name}{name of target table.}
}
\value{
A data frame on success, or NULL.
}
\description{
Generic function to get one dataset from stock_db.
}
\section{Methods (by class)}{
\itemize{
\item \code{gta_db}: get a table dataset from a database of gta_db class
}}

\examples{
ds_trd_mnth.df <- get_table_dataset(stock_db, table_name = "TRD_Mnth_月个股回报率")
}
