% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-dates.R
\name{guess_dates_period}
\alias{guess_dates_period}
\title{Guess periodicity of date series}
\usage{
guess_dates_period(dates_series, regular = FALSE)
}
\arguments{
\item{dates_series}{A vector of date/timestamps.}

\item{regular}{Whether to test date series as a regular periodic series
or not. Since irregular test is looser than regluar test, irregluar test
on regluar series will return true. Default is FALSE, which means we will
use looser test(irregular).}
}
\value{
return "D" for daily, "M" for monthly, "Q" for quarterly, "Y" for yearly,
 "U" for unkown.
}
\description{
Guess periodicity of date series
}
\seealso{
Other utility functions of dates: \code{\link{is_periodic_dates}}
}
\concept{utility functions of dates}
