% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-cliui.R
\name{help_fun}
\alias{help_fun}
\alias{help_fun_}
\title{Display usage of function}
\usage{
help_fun(fun, argument_desc = NULL, examples = NULL)

help_fun_(quo_fun, argument_desc = NULL, examples = NULL)
}
\arguments{
\item{fun}{a function to be display usage.}

\item{argument_desc}{a character arguments description.}

\item{examples}{a character examples.}

\item{quo_fun}{a function quoted by \code{\link[rlang:quo]{rlang::quo()}} or \code{\link[rlang:enquo]{rlang::enquo()}} to be
display usage.}
}
\value{
standard output
}
\description{
Display help for a function by arguments usage and examples.
}
\seealso{
Other utils_cliui: 
\code{\link{interactive_call}()},
\code{\link{interactive_fun}()}
}
\concept{utils_cliui}
