% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock-db.R, R/stock-db-gta.R
\docType{methods}
\name{init_stock_db}
\alias{init_stock_db}
\alias{init_stock_db,gta_db-method}
\title{Init param of stock db}
\usage{
init_stock_db(stock_db, ...)

\S4method{init_stock_db}{gta_db}(stock_db, ...)
}
\arguments{
\item{stock_db}{A stock database object to operate.}
}
\value{
TRUE if success, else FALSE.
}
\description{
Generic function to initiate params of a stock database.
}
\section{Methods (by class)}{
\itemize{
\item \code{gta_db}: Init param of database of gta_db class
}}

\examples{
stock_db <- stock_db(gta_db, "GTA_SQLData")
open_stock_db(stock_db)
init_stock_db(stock_db)
}
\seealso{
Other stock_db generics: \code{\link{close_stock_db}},
  \code{\link{code2name}},
  \code{\link{fetch_table_dataset}},
  \code{\link{get_factor_indicator}},
  \code{\link{get_factors_info}},
  \code{\link{get_market_return}},
  \code{\link{get_stock_dataset}},
  \code{\link{get_stock_return}},
  \code{\link{get_table_dataset}},
  \code{\link{list_stock_tables}}, \code{\link{name2code}},
  \code{\link{open_stock_db}}
}
