% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock-db.R, R/stock-db-gta.R
\name{init_stock_db}
\alias{init_stock_db}
\alias{init_stock_db,gta_db-method}
\title{Init param of stock db}
\usage{
init_stock_db(stock_db, ...)

\S4method{init_stock_db}{gta_db}(stock_db, ...)
}
\arguments{
\item{stock_db}{A stock database object to operate.}

\item{...}{Extra arguments to be passed to methods.}
}
\value{
TRUE if succeed else FALSE.
}
\description{
Generic function to initiate params of a stock database.
}
\section{Functions}{
\itemize{
\item \code{init_stock_db,gta_db-method}: Init param of database of gta_db class
}}

\examples{
\dontrun{
stock_db <- stock_db(gta_db, "GTA_SQLData")
open_stock_db(stock_db)
init_stock_db(stock_db)
}

}
\seealso{
Other stock_db generics: 
\code{\link{close_stock_db}()},
\code{\link{code2name}()},
\code{\link{dir_path_db}()},
\code{\link{fetch_table_dataset}()},
\code{\link{get_factors_info}()},
\code{\link{get_factors}()},
\code{\link{get_financial_report}()},
\code{\link{get_indicators_from_source}()},
\code{\link{get_indicators}()},
\code{\link{get_market_return}()},
\code{\link{get_profile}()},
\code{\link{get_riskfree_rate}()},
\code{\link{get_spt_stocks}()},
\code{\link{get_stock_dataset}()},
\code{\link{get_stock_industry}()},
\code{\link{get_stock_info}()},
\code{\link{get_stock_return}()},
\code{\link{get_table_dataset}()},
\code{\link{list_stock_tables}()},
\code{\link{name2code}()},
\code{\link{open_stock_db}()},
\code{\link{save_indicators_to_source}()}
}
\concept{stock_db generics}
