% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock-db.R, R/stock-db-gta.R
\docType{methods}
\name{name2code}
\alias{name2code}
\alias{name2code,gta_db-method}
\title{Translate name into code}
\usage{
name2code(x, name, ...)

\S4method{name2code}{gta_db}(x, name, ...)
}
\arguments{
\item{x}{A object containg code/name infomation}

\item{name}{A name or a vector of names to be translated}

\item{...}{Other arguments to be provided to underlyling functions}
}
\value{
A code or a vector of codes
}
\description{
Generic function to translate name into code
}
\section{Methods (by class)}{
\itemize{
\item \code{gta_db}: Translate name into code in a database of gta_db class
}}

\seealso{
Other stock_db generics: \code{\link{close_stock_db}},
  \code{\link{code2name}},
  \code{\link{fetch_table_dataset}},
  \code{\link{get_factor_indicator}},
  \code{\link{get_factors_info}},
  \code{\link{get_market_return}},
  \code{\link{get_stock_dataset}},
  \code{\link{get_stock_return}},
  \code{\link{get_table_dataset}},
  \code{\link{init_stock_db}},
  \code{\link{list_stock_tables}},
  \code{\link{open_stock_db}}
}
