% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock-db-gta.R, R/stock-db.R
\name{name2code.gta_db}
\alias{name2code.gta_db}
\alias{name2code}
\alias{name2code.code_name_list}
\title{Translate name into code}
\usage{
\method{name2code}{gta_db}(stock_db, name, type = c("stock", "field"))

name2code(x, name, ...)

\method{name2code}{code_name_list}(x, name)
}
\arguments{
\item{name}{a name or a vector of names to be translated}

\item{x}{a object containg code/name infomation}

\item{...}{other params to be provided to underlyling functions}
}
\value{
a code or a vector of codes
}
\description{
Generic function to translate name into code
}
\section{Methods (by class)}{
\itemize{
\item \code{gta_db}: Translate name into code in a database of gta_db class

\item \code{code_name_list}: Translate name into code in a object of code_name_list
}}

\examples{

}
