% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock-db.R, R/code-name-list.R,
%   R/stock-db-gta.R
\name{name2code}
\alias{name2code}
\alias{name2code,code_name_list-method}
\alias{name2code,gta_db-method}
\title{Translate name into code}
\usage{
name2code(x, name, exact_match = TRUE, ...)

\S4method{name2code}{code_name_list}(x, name, exact_match = TRUE, ...)

\S4method{name2code}{gta_db}(
  x,
  name,
  exact_match = TRUE,
  type = c("stock", "field", "industry", "factor", "indicator"),
  ...
)
}
\arguments{
\item{x}{A object containing code/name information.}

\item{name}{A name or a vector of names to be translated.
if exact_match is FALSE, name could be regular expression for
non-exact matching.}

\item{exact_match}{A logic to determine use exact matching or not.
Default TRUE means to use exact matching}

\item{...}{Extra arguments to be passed to methods.}

\item{type}{A character of conversion type, e.g, "stock", "field", "industry".
Default "stock" means to covert stock name to stock code.}
}
\value{
\preformatted{ A code or a vector of codes.
}
}
\description{
Generic function to translate name into code
}
\section{Functions}{
\itemize{
\item \code{name2code,code_name_list-method}: Translate name into code in a object of code_name_list

\item \code{name2code,gta_db-method}: Translate name into code in a database of gta_db class
}}

\seealso{
Other stock_db generics: 
\code{\link{close_stock_db}()},
\code{\link{code2name}()},
\code{\link{dir_path_db}()},
\code{\link{get_factors_info}()},
\code{\link{get_factors}()},
\code{\link{get_financial_report}()},
\code{\link{get_indicators_from_source}()},
\code{\link{get_indicators_info}()},
\code{\link{get_indicators}()},
\code{\link{get_industry_info}()},
\code{\link{get_market_return}()},
\code{\link{get_profile}()},
\code{\link{get_riskfree_rate}()},
\code{\link{get_spt_stocks}()},
\code{\link{get_stock_dataset}()},
\code{\link{get_stock_industry}()},
\code{\link{get_stock_info}()},
\code{\link{get_stock_return}()},
\code{\link{get_table_dataset}()},
\code{\link{init_stock_db}()},
\code{\link{list_stock_tables}()},
\code{\link{open_stock_db}()},
\code{\link{save_indicators_to_source}()}
}
\concept{stock_db generics}
