% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor-normalize.R
\name{normalize_factors}
\alias{normalize_factors}
\title{Normalize the factors dataset}
\usage{
normalize_factors(ds_factors, factors_list = NULL, group_by = NULL,
  clean_extremes_method = c("sigma", "mad"),
  standard_method = c("normal", "rank"), ...)
}
\arguments{
\item{ds_factors}{a factors dataset.}

\item{group_by}{a character vector of fields as group data for scaling, NULL
is default value which means no group settting.}

\item{clean_extremes_method}{method of cleaning extremes befre standardization
, i.e. "sigma","mad", by default "sigma".
see details: \code{\link{clean_extremes_sigma}}, \code{\link{clean_extremes_mad}}}

\item{standard_method}{method of standardizing data, i.e. "normal","rank",
by default "normal".
see details: \code{\link{standardize_normal_scale}}, \code{\link{standardize_rank_scale}}}

\item{...}{additional arguments to clean_extremes_method and standard_method.}
}
\value{
a dataset of normalized factors
}
\description{
Normalze factors dataset by cleaning extremes and standardation.
}
