% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock-db.R, R/stock-db-gta.R
\docType{methods}
\name{open_stock_db}
\alias{open_stock_db}
\alias{open_stock_db,gta_db-method}
\title{Open stock database}
\usage{
open_stock_db(stock_db, ...)

\S4method{open_stock_db}{gta_db}(stock_db, ...)
}
\arguments{
\item{stock_db}{A stock database object to operate}
}
\value{
TRUE if succeed otherwise FALSE
}
\description{
Generic function to open a stock database
}
\section{Methods (by class)}{
\itemize{
\item \code{gta_db}: Open a database of gta_db class
}}

\examples{
\dontrun{
library(zstmodelr)
stock_db <- stock_db(gta_db, "GTA_SQLData")
open_stock_db(stock_db)
close_stock_db(stock_db)
}
}
\seealso{
Other stock_db generics: \code{\link{close_stock_db}},
  \code{\link{code2name}}, \code{\link{dir_path_db}},
  \code{\link{fetch_table_dataset}},
  \code{\link{get_factors_info}},
  \code{\link{get_factors}},
  \code{\link{get_finacial_report}},
  \code{\link{get_indicators_from_source}},
  \code{\link{get_indicators}},
  \code{\link{get_market_return}},
  \code{\link{get_profile}},
  \code{\link{get_riskfree_rate}},
  \code{\link{get_spt_stocks}},
  \code{\link{get_stock_dataset}},
  \code{\link{get_stock_industry}},
  \code{\link{get_stock_return}},
  \code{\link{get_table_dataset}},
  \code{\link{init_stock_db}},
  \code{\link{list_stock_tables}}, \code{\link{name2code}},
  \code{\link{save_indicators_to_source}}
}
\concept{stock_db generics}
