% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicator-define.R, R/indicator-define-gta.R
\docType{methods}
\name{parse_indicator_vars}
\alias{parse_indicator_vars}
\alias{parse_indicator_vars,gta_db-method}
\title{Parse vars in indicator expr from stock_db}
\usage{
parse_indicator_vars(stock_db, indicator_expr, ...)

\S4method{parse_indicator_vars}{gta_db}(stock_db, indicator_expr, ...)
}
\arguments{
\item{stock_db}{A stock database object to operate.}

\item{indicator_expr}{A expr of indicator to parase.}
}
\value{
A character vector of name of vars if succeed, otherwise NULL.
}
\description{
Generic function to parse vars in indicator expr from stock_db.
}
\section{Methods (by class)}{
\itemize{
\item \code{gta_db}: parse vars in indicator expr from
database of gta_db class.
}}

\seealso{
Other indicator define functions: \code{\link{create_attribute_def_fun}},
  \code{\link{create_indicator_def_fun}},
  \code{\link{get_indicator_defs}},
  \code{\link{get_indicator_vars}},
  \code{\link{ind_attr_def_indcd}},
  \code{\link{ind_attr_def_trdstat}},
  \code{\link{prioritize_indicator_defs}}
}
\concept{indicator define functions}
