% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-mgt.R
\name{read_import_file}
\alias{read_import_file}
\title{Read a raw file for importing to stock db}
\usage{
read_import_file(
  input_file,
  input_template = NULL,
  input_type = c("csv", "txt"),
  input_dir = NULL,
  start_index = 2L,
  ignore_problems = TRUE,
  log_dir = "./log"
)
}
\arguments{
\item{input_file}{A character of name or path of input data file,
which could use regular expression to read multiple files into a combined
dataframe, e.g. \code{test[0-9]*.txt}" will read files like test.txt, test1.txt,...,
test9.txt..., etc., the files ware queued in ascending alphabet order.}

\item{input_template}{A character of name of input template file.
Default NULL means use as input_file when reading single files or
the first file in input_file when reading multiple files.}

\item{input_type}{A character of input file format, e.g. "txt", "csv".}

\item{input_dir}{A path of working dir of input file, if NULL, use dir of
input file as working dir, default NULL.}

\item{start_index}{A integer of start index of first line of actual records
in data file, by default 2L, which means first line is header and actual
data starts from second line.}

\item{ignore_problems}{A logic flag to determine whether to ignore problems
when covert data, if TRUE, continue to import data but log problems into
log file, otherwise abort reading process.  By default TRUE.}

\item{log_dir}{Path of log dir for saving problem log file, by
default"./log", if the log path doesn't existed, it  will be created.}
}
\value{
A dataframe of raw data for importing if succeed, otherwise NULL.
}
\description{
Read a raw file into a dataframe for processing or importing.
}
\seealso{
Other data management: 
\code{\link{clear_tables}()},
\code{\link{convert_import_file}()},
\code{\link{get_datasource}()},
\code{\link{import_table}()},
\code{\link{process_files}()},
\code{\link{ttm_financial_report}()},
\code{\link{update_db}()},
\code{\link{write_import_file}()}
}
\concept{data management}
