% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock-db.R
\name{stock_db}
\alias{stock_db}
\title{subclass object factory of stock_db class}
\usage{
stock_db(stock_db_class, ...)
}
\arguments{
\item{stock_db_class}{Specific creator of stock database class,
e.g. gta_db}

\item{...}{Addition params used by specific class of stock database
(e.g. a dsn string "GTA_SQLData").}
}
\value{
A object of stock db.
}
\description{
factory for creating for subclass object of stock_db class
}
\examples{
\dontrun{
library(zstmodelr)
stock_db <- stock_db(gta_db, "GTA_SQLData")
open_stock_db(stock_db)
close_stock_db(stock_db)
}
}
