% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-timeseries.R
\name{ts_asfreq}
\alias{ts_asfreq}
\alias{ts_asfreq.tbl_df}
\alias{ts_asfreq.timeSeries}
\title{Utility Frequency Conversion Method of Setting Frequncey of Timeseries}
\usage{
ts_asfreq(ts_dataset, freq_rule = c("day", "month", "quarter"),
  fillna_method = c("nfill", "ffill", "bfill"), ...)

\method{ts_asfreq}{tbl_df}(ts_dataset, freq_rule = c("day", "month",
  "quarter"), fillna_method = c("nfill", "ffill", "bfill"),
  date_index_field = c("date"), by_group = NULL)

\method{ts_asfreq}{timeSeries}(ts_dataset, freq_rule = c("day", "month",
  "quarter"), fillna_method = c("nfill", "ffill", "bfill"))
}
\arguments{
\item{ts_dataset}{a timeseries of tibble/timeSeries.#}

\item{freq_rule, }{the offset string or object representing target conversion,
e.g. "Day", "Month", "Quarter", default "Day".}

\item{fillna_method, }{method to fill holes in reindexed Series, e.g.
"nfill", "bfill","ffill", default nfill(fiil with NA).}

\item{date_index_field}{the name of date index field of ts_df for resample,
default 'date', Column must be date-like.
Only be used for tibble dataset.}

\item{by_group}{a character vector of fields as group data for asfreq.
Only be used for tibble dataset.}
}
\value{
a converted timeseres
}
\description{
Convert timeseries at specified frequency by setting new frequency of data index
}
\details{
Convert timeseries to specified frequency by refrequencing date index.
Optionally provide filling method to pad/backfill missing values.
Return originaldata conformed to a new index with the specified frequency.

ts_asfreq is more appropriate if use original the data at the new frequency.
}
\section{Methods (by class)}{
\itemize{
\item \code{tbl_df}: Set new frequency for timeseries of tibble dataset

\item \code{timeSeries}: Set new frequency for timeseries of timeSeries dataset
}}

