% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-timeseries.R
\name{ts_asfreq}
\alias{ts_asfreq}
\alias{ts_asfreq.tbl_df}
\alias{ts_asfreq.timeSeries}
\title{Frequency Conversion Function of Setting Frequency of timeseries}
\usage{
ts_asfreq(
  ts_dataset,
  freq_rule = c("day", "month", "quarter", "year"),
  fillna_method = c("nfill", "ffill", "bfill"),
  ...
)

\method{ts_asfreq}{tbl_df}(
  ts_dataset,
  freq_rule = c("day", "month", "quarter", "year"),
  fillna_method = c("nfill", "ffill", "bfill"),
  ...,
  date_index_field = c("date"),
  key_fields = NULL,
  parallel = TRUE
)

\method{ts_asfreq}{timeSeries}(
  ts_dataset,
  freq_rule = c("day", "month", "quarter"),
  fillna_method = c("nfill", "ffill", "bfill"),
  ...
)
}
\arguments{
\item{ts_dataset}{A timeseries of tibble/timeseries.#}

\item{freq_rule}{The offset string or object representing target conversion,
e.g. "day", "month", "quarter", "year", default "day".}

\item{fillna_method}{Method to fill holes in reindexed Series, e.g.
"nfill", "bfill","ffill", default nfill(fill with NA).}

\item{...}{Extra arguments to be passed to methods.}

\item{date_index_field}{Name of date index field of ts_df for refrequecy,
default 'date', Column must be date-like. Only be used for tibble dataset.}

\item{key_fields}{A character vector of key fields, which identify unique
observation in each date. Only be used for tibble dataset.}

\item{parallel}{A logic to determine whether to use parallel processing.
Only be used for tibble dataset.}
}
\value{
\preformatted{       A converted timeseres
}
}
\description{
Generic function tp convert timeseries at specified frequency by setting new frequency of data index
}
\details{
Convert timeseries to specified frequency by refrequencing date index.
Optionally provide filling method to pad/backfill missing values.
Return original data conformed to a new index with the specified frequency.

ts_asfreq is more appropriate if use original the data at the new frequency.
}
\section{Methods (by class)}{
\itemize{
\item \code{tbl_df}: Set new frequency for timeseries of tibble dataset

\item \code{timeSeries}: Set new frequency for timeSeries of timeSeries dataset
}}

\seealso{
Other utils_timeseries: 
\code{\link{ts_lag}()},
\code{\link{ts_resample}()}
}
\concept{utils_timeseries}
