% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-timeseries.R
\name{ts_resample}
\alias{ts_resample}
\alias{ts_resample.tbl_df}
\alias{ts_resample.timeSeries}
\title{Utility Frequency Conversion Method of Resamping Timeseries}
\usage{
ts_resample(ts_dataset, freq_rule = c("day", "month", "quarter"),
  fillna_method = c("nfill", "ffill", "bfill"), agg_fun = mean, ...)

\method{ts_resample}{tbl_df}(ts_dataset, freq_rule = c("day", "month",
  "quarter"), fillna_method = c("nfill", "ffill", "bfill"),
  agg_fun = mean, ..., date_index_field = c("date"),
  key_fields = NULL, parallel = TRUE)

\method{ts_resample}{timeSeries}(ts_dataset, freq_rule = c("day",
  "month", "quarter"), fillna_method = c("nfill", "ffill", "bfill"),
  agg_fun = c("mean", "sum"), ...)
}
\arguments{
\item{ts_dataset}{A timeseries of tibble or timeSeries.}

\item{freq_rule}{The offset string or object representing target conversion,
e.g. "day", "month", "quarter", default "Day".}

\item{fillna_method}{method to fill holes in reindexed Series, e.g.
"nfill", "bfill","ffill", default nafill(fill NA)}

\item{agg_fun}{Function to aggregate values of group data for new timestamp,
default setting is mean}

\item{...}{Argments passed to agg_fun}

\item{date_index_field}{Name of date index field of ts_df for resample,
default 'date', Column must be date-like.
Only be used for tibble dataset.}

\item{key_fields}{A character vector of key fields, which identify unique
observation in each date. Only be used for tibble dataset.}

\item{parallel}{A logic to deterimine whether to use parallel processing.}
}
\value{
A converted timeseres.
}
\description{
Convert timeseries to specified frequency by resamploing date index
}
\details{
Convert timeseries to specified frequency by resampling date index.
Optionally provide filling method to pad/backfill missing values.
Return aggregating data conformed to a new index with the specified frequency.

resample is more appropriate if an operation, such as summarization, is
necessary to represent the data at the new frequency.
}
\section{Methods (by class)}{
\itemize{
\item \code{tbl_df}: Resamping timeseries of tibble dataset

\item \code{timeSeries}: Resamping timeseries of timeSeries dataset
}}

