% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-mgt.R, R/data-mgt-gta.R
\docType{methods}
\name{update_db}
\alias{update_db}
\alias{update_db,gta_db-method}
\title{Update tables in stock database with new data}
\usage{
update_db(stock_db, data_source = get_datasource(stock_db),
  retry_log = NULL, log_file_prefix = "update_db_log",
  log_dir = "./log", ...)

\S4method{update_db}{gta_db}(stock_db,
  data_source = get_datasource(stock_db), retry_log = NULL,
  log_file_prefix = "update_db_log", log_dir = "./log", ...)
}
\arguments{
\item{stock_db}{A stock database object to operate.}

\item{data_source}{Data source info produced by get_datasource(stock_db)}

\item{retry_log}{Log file for re-importing tables with recording failure
in log file. If NULL, it will upate all talbles in data_source, otherse
only update on these tables with recording failure in log file.
By default NULL.}

\item{log_file_prefix}{A character of log file prefix to name log file.
  Log file is named as format of "log_file_prefix_XXXX_(current).csv"
Default is "update_db_log".}

\item{log_dir}{Path to save updating log file. NULL means to use "./" as
log dir. Default "./log".}
}
\value{
NULL invisibly. Raise error if anything goes wrong.
}
\description{
Generic function to update tables in stock database by importing new data
}
\details{
By combining infomation from data_source and re-try log file , it update all
tables from raw data of datasource in following steps:

\enumerate{
   \item Build target tables info from all tables in datasource or some
   tables with recoding failure in retry log file;
   \item Update target tables by importing data from datasoure;
   \item Save update log, etc.
}

Update log is saved in log dir, like "update_log_XXXX(current).csv"
}
\section{Methods (by class)}{
\itemize{
\item \code{gta_db}: update tables in a database of gta_db class
}}

\seealso{
Other data managment: \code{\link{clear_tables}},
  \code{\link{convert_import_file}},
  \code{\link{get_datasource}}, \code{\link{import_table}},
  \code{\link{process_files}},
  \code{\link{read_import_file}},
  \code{\link{ttm_financial_report}},
  \code{\link{write_import_file}}
}
\concept{data managment}
