% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zstmodelr-package.R
\docType{package}
\name{zstmodelr-package}
\alias{zstmodelr}
\alias{zstmodelr-package}
\title{zstmodelr: Stock Factor Models for China Stock Market Investment}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

An implementation of analytical tools of stock factors models for 
   investing China stock market.
}
\details{
The main goals of zstmodelr is to leverage traditional financial theory and
modern ML method to build stock models for China stock market.

To learn more about zstmodelr, start with the vignettes:
\code{browseVignettes(package = "zstmodelr")}
}
\section{Options}{

\itemize{
\item \code{testthat.use_colours}: Should the output be coloured? (Default: \code{TRUE}).
\item \code{testthat.summary.max_reports}: The maximum number of detailed test
reports printed for the summary reporter (default: 10).
\item \code{testthat.summary.omit_dots}: Omit progress dots in the summary reporter
(default: \code{FALSE}).
}
}

\examples{
\dontrun{
library(testthat)
a <- 9
expect_that(a, is_less_than(10))
expect_lt(a, 10)
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/chriszheng2016/zstmodelr}
  \item Report bugs at \url{https://github.com/chriszheng2016/zstmodelr/issues}
}

}
\author{
\strong{Maintainer}: Chris Zheng \email{chriszheng@vip.sina.com.cn} (\href{https://orcid.org/CHRISZHENG'S-ORCID-ID}{ORCID})

}
\keyword{internal}
